Add file list of files/directories in tools/base to run gradle tests.
We are moving to a unified bazel PSQ. We are not ready to run all the
tools/base changes in bazel. Thus, we run the non tools/base changes
with bazel and then if the change is tools/base, and the file or
directory is listed in this file, we will run the gradle tests for those
CL's. Otherwise, we only run the bazel build/test. As we implement
tools/base tests in bazel, we can remove stuff from this file, until
there is nothing left.

Test: Run local psq master and slave

Change-Id: I4ce12e08d41a0dd06d7146751d17e211874f1ac6