Do not allow FileCache to lock output files.
When FileCache copies a cached file to a requested output location,
it currently locks both the cached file and the output location to
prevent concurrent access to them from multiple threads/processes.

Locking the cached files is required, but locking the output files is
optional. With this commit, we no longer let FileCache lock output
files.

This will help to reduce the performance overhead due to file locking
(such as the one reported in http://b.android.com/231831).

As a result of this change, when the clients use the cache to create an
output file outside the cache directory, they need to make sure that
multiple threads/processes do not create the same output file, or they
need to have a way to manage that situation.

Test: New test in FileCacheTest
Change-Id: I4945b79d65be6ceccfcb6168d931a48aaf28b19a