Separate artifact query for classes and res.
In order to have the finest grained dependency between projects, we
can separate the generation of java resources and java classes in the
android library projects (this is already done in the previous
commit), however because pure java projects were still publishing a
single jars, it was difficult to actually make the java resources
task consume only the resource jars coming from library projects.

In the newer Gradle nightly, the java plugin now publishes its
intermediate as CLASS_DIRECTORY and RESOURCES_DIRECTORY (these
names are temporary and shouldn't mean that it's only a dir that's
published).

So this CL now makes the classes and resources of the android lib
module be published with the same types and the Transform streams
now properly match the ContentType with the right artifact type.

This will reduce instances where a CLASS change in a library project
trigger java resource merging in the app project.

Because the query method in VariantScope were getting out of
control and their name didn't quite explain what they queried, this
change also replaces all the method with a single method (well 2
actually: one for FileCollection and one for ArtifactCollection),
and make the caller explicitely declare which config to use
(compile vs package), which scope (all, external, sub-projects),
and what artifact type to query for.

This change also fixes an issue with the provided scope of streams
which still queried the Variantconfiguration for data which is
incorrect.

Finally, this change also fixes the filtering of local jars for
external libraries which removed all file-based jars, even
transitive ones, instead of just the ones for the current module.

Test: existing tests.

Change-Id: Ibcbba70eb484777259ed789e5bdcef16bf21078f