Deletes code not used in the profiler plugin.
This code is not used and was making the instrumentation more expensive and in two
passes. More over there is a bigger change coming that will probably require
making our plugin have no dependencies (including guava or asm) to not interfere
with the android gradle plugin, and there is no point in updating all this code.

Change-Id: Ifaad39c97d0544393e3d7553343de953a83dab49
Test: Will write an integration test once the feature is finished.