Building dex archives
Adding conversion of the .class file inputs to the
dex archives. This is done by extracting 2 methods
that do the actual conversion from dx, and directly
invoking them.

Class inputs are supported from directories and jars.
Dex archives can be created in a directory, or a jar.

Conversion is happening class per class, and after the
conversion, the dex files is written to the archive.

Test: ClassFileInputTest, DexArchiveBuilderTest
Change-Id: If1d6a897dca7c40bf7fbd0e0a2787d1afdedb643