Change the way intermediate streams work.
1. It used to be based on a structure of folders, with
the name of the input/output as a filename but this
prevents safely using names that may contain non
path safe character.

Move to a flat directory structure with an index
(json file) containing the content and the meta-data
(format, scopes, content types, etc...)

Based on this the OriginalStream can now receive
an ArtifactCollection, and ensure that the name of
each Folder/Jar Input that it creates is unique, and
based on the address of the item.

2. In order to ensure that ArtifactCollections returned
by VariantScope contains all that is needed, refactored
the getArtifactFileCollection and the getArtifactCollection
methods to always do exactly the same thing, especially
when it comes to handling test artifacts and duplicated
dependencies for test app.

This required adding the ability to add/remove from an
ArtifactCollection which is not supported out of the box.
This means adding new custom implementations of the interface.

In turn this allows simplifying a few tasks that used
ArtifactCollection and a manual handling of the tested
artifacts.

3. Fixed a bunch of tests, including unit tests and integration
ones. In particular integration tests that look into
the transform output folder must now be able to read
the json index file. There's a new class (TransformOutputContent)
that allows to do it and return the same SubStream objects
used internally. New Truth Subjects for both SubStream
and TransformOutputContent were also added.

Test: existing + fixing a bunch.
Change-Id: I0612e20362ec9a7bfe4fd07ac81ee8def99fd1b9