Make basic test project ready for AAPT2
Modifies the AndroidManifest in basic test project by removing the
duplicate permission.
In the case when in the manifest two or more permissions with the same
last token in the name are declared, AAPT1 would overwrite all with the
last one in Manifest.java. AAPT2 though will try to create constants for
each of them, therefore leading to duplicate resource name conflicts in
the Manifest.java.
Until AAPT2 becomes more forgiving, we need to remove duplicate
permissions.

Test: BasicTest, BasicTest2 with enableAapt2=true
Change-Id: I4f70e60667f79596a24377a699812242684d9541