Use build session variables for correct life time and scope.
This commit introduces the BuildSessionVariable class to address two
potential issues when using static variables:

1. Life time: Static variables persist across builds since the same
   Gradle daemon can be used to execute multiple builds. This is usually
   unintended, as we want to limit the variables' life time to a build.
   For example, if a build fails or is canceled by Ctrl-C, the states of
   static variables may be corrupted and may break the next build.

   Even if the build completes normally, the states of static variables
   are not "fresh" (e.g., a counter may have increased and the next
   build's counter will not start from 0).

2. Scope: Static variables are unique per class loader of the class that
   loads them. If the plugin is loaded multiple times by different class
   loaders, static variables are no longer unique. This can be addressed
   by using JVM-wide variables; however, JVM-wide variables do not
   handle the issue with the variables' life time, whereas build session
   variables can handle both issues.

A build session variable is similar to a JVM-wide variable, with one
important extension: Its life time is limited to a build.

Specifically, the variable is initialized with a default value when the
variable is first created and will be recreated and re-initialized at
the start of every subsequent build within the same Gradle daemon.

This commit modifies the build cache to use build session variables. We
should also consider using build session variables for some other
mutable static fields in our code that have the above two issues.

Bug: 37140936

Test: BuilSessionVariableTest, JvmWideVariableTest; Manually
upgrade/downgrade a project between 2.3.0 and the latest plugin version
and check that building/syncing succeeds

Change-Id: Ia551db42336560bafedf62e3410d49650dd6e115