Using new dx with min sdk
When invoking dx we will always pass the
min-sdk-version which is required in order
to support 24+ (default/static methods)
and 26+ (invoke-lambda) features.

This CL adds minSdkVersion parameter whenever we
are invoking dx. Also, it has been added as an input
to the affected tasks, and as part of the cache key.

Test: AntennaPodInstantRunTest, DesugarAppTest
Change-Id: Ia6bf1052d74afd9c5310d0ef040d8f5306cfdb1e