Fix NPE in MergeOutputWriters.
The remove() method in MergeOutputWriters currently will throw a
NullPointerException if it is called on a file that either doesn't exist
or is a directory.

The root cause is at IncrementalRelativeFileSets.fromZip() which
currently might accidentally add directories to the inputs, which must
only include files.

This is the issue filed in http://b/37324898 and one of the two issues
found in http://b/37324855.

Test: IncrementalRelativeFileSetsTest
Bug: http://b/37324898, http://b/37324855
Change-Id: I85689ac1e94ce2f293d64fdaeb7d21960986ba58