Make AAPT2 generate R.txt
While AAPT2 wasn't generating R.txt we used a workaround of making AAPT2
generate copies of R.java files for the libraries the current module
depended on.

Now, as the flag for R.txt generation was added, we can go back to
relying on R.txt and generating the filtered R.java files ourselves,
just as we did with AAPT1.

This CL also disables non-JNI AAPT2 scenario for MergeResourcesTest as
non-JNI AAPT2 is missing all the new flags we have in the JNI AAPT2.

Test: MergeResourcesTest
Change-Id: I56959deb15818b0a1f9f7f6ae83be8c0b5798c04