Remove dontpreverify flag from default rules
This CL removes the dontpreverify flag from
the default Proguard rules shipped with the plugin.
Proguard would not update the StackMapTable when
dontpreverify is specified, which could lead to
JVM VerificationError for any tests run on the host.
Also, any published aar could not be used to run
tests locally on the host JVM. Performance benchmarks
using the santa-tracker project did not show any
performance impact.

BUG=37137716
Test: existing

Change-Id: I18b9b1c1c19772c8bfb82f06ac5c808e9995c630