[AIA] Use protobuf to read instant app sdk
In AIA SDK, we have both an .xml and a .pb files describing the needed
metadata for the moment. Studio consumes the xml. We want to use only
protos.

This CL adds the support for reading from the pb file. It tries first to
get the metadata from a proto file and if it fails, it gets from the
xml.

The jars describing the proto structue are generated from the .proto
file in the internal AIA repository and jarjar'ed so there is no
conflict between proto2 and proto3. This CL implements a gradle script
to generate the jar from the .proto file.

Test: unit tests.
Change-Id: Icb7c824f45ee87d8af1edd446c1d1e4d549c42f9