Use FileCollection for data binding dependency artifacts.
This creates a new task output for the result of the
merge, and use it to feed the javactask (which in turns
properly sets up the task dependency).

This also fixes the published data binding artifacts for
aar that dependended on the merge instead of the javac task
which generates it via the apt.

Finally, exposes the merge output as a file collection via
the variant API so that Kotlin can depend on it.

Bug: 62206310
Change-Id: Ieae5242dbee405c7b94e12ac6cbd752b4f27124f
Test: existing tests.