Prevent XXE attacks
Prevent XML External Entity attacks by refusing to load external
DOCTYPE declarations. For most XML parsers we disallow DOCTYPE
declarations altogether (for example, in device definition files
there will never be DOCTYPE declarations). For some parsers,
such as the one used to process XML files encountered by lint
and by the SVG importer, we allow DOCTYPE declarations to be
present in the document, but we don't process them (the resource
is not loaded etc).

There are tests which check this.

Test: Covered by existing tests (plus new tests added.)

Change-Id: I2f1978bc5458ba2b2b2d6ffbc9df5710c487a4e4