Better error message in case of variant aware dep failure.
In case there is a build-type/flavor mismatch we need to provide
a better error during sync with the IDE. Right now the error does
not include any details besides "failed to resolve deps".

However to do this we need to support longer error messages in the
sync issue. This CLs adds getMultiLineMessage as an optional longer
message to the sync issue. The IDE can display the shorter message
in the message window and show the full text with a "More..." link.

Change-Id: Ib7c18d20a78920e975bdb556465a2bcffafb1123
Test: new test for the attr mismatch sync error.
Bug: 63545042