change merge logs file format to be more efficient.
the merge logs stored 1:1 entries between source and output files.
Instead flatten to 1:n entries as well as using a collection of mappings
instead of storing each mapping individually.

this lead to 90% file size reduction, large projects used to generate
up to 30 mb of logs, now it's down to less than 1.
this is a step forward better supporting large projects with lots of
resources.

we support reading all file formats for backward compatibility.

Test: existing + adding new ones.
Bug: 62949797
Change-Id: I765c0783f660a1cd2811e7440ec3e8b883661dd8