Move VariantConfiguration to gradle-core.
This allows us to start refactoring the ProductFlavor/BuildType
DSL to not extend the model versions of these interfaces.

Without this change, we'd need to make builder depend on
gradle-api which is not great.

Also, the goal of the new Variant API is to merge
VariantScope, VariantData and (Gradle)VariantConfiguration
into a single object anyway, so this goes in the right
direction.

Change-Id: I16c2a842e9dd67921690f30ee11ac41887dc1499
Test: existing integ tests + moved unit tests and fixtures.