Refactoring of ErrorReporter
Main refactorings:
SyncIssueHandlder -> EvalIssueReporter
ErrorReporter -> ConfigurableErrorHandler
(new) DeprecationReporter

Also made EvalIssueReporter only about reporting
(moving hasSyncIssue to ConfigErrorHandler).

In most (all?) places that needs to do reporting,
the var/field is now EvalIssueReporter instead of
ErrorReporter to only give access to the reporting
API.

New DeprecationReporter to provide a more targeted
API for it. The API requires an enum which is when
the deprecation is targeted to be deleted. (this
will allows us to search for usage of the enums and
delete all warnings as well as DSL methods & properties.)

Moved a few usage of handleSyncIssue to the deprecation
reporter.

Finally, also removed the errorReporter from a bunch
of classes. We have a few god objects that have it
anyway so it seems better than having every variantscope,
variantData, API objects etc... hold it.

Test: existing tests with some fixes.
Change-Id: I9aef32ee6b965b0c353d7979a7d38a462dc2ba02