Change name of directory that contains merging log files.
Previously, these log files used to be kept under the
"intermediates/blame/../multi" directory, and we did not clean it at the
start of a full build.

Recently, we introduced a new format for these log files (commit
3b7ab4fe2cce3c75d81a7d63dd90bc01dfac8f5c). Therefore, if the users
switch between versions of the plugin that have different formats of the
log files, the next build will fail to read the existing log files. (We
tried to support a newer plugin version reading the old format, but we
still cannot support an older plugin version reading the new format.)

To work around this issue, we need to use a different directory called
"intermediates/blame/../multi-v2" to contain log files with the new
format.

Note that, with a recent change (commit
4e603b3ba720429317c854da8c69d6b9a8828270), we now also clean these log
files at the start of a full build. Therefore, moving forward, we won't
need to use a new directory again even if we introduce yet another new
format.

Also, as it is no longer necessary to support reading the old format,
this commit removes that part of the code as well.

Bug: 65263212
Test: Updated MergingLogTest
Change-Id: Ia5837b940d10a7a8b3686029b4986cce98093771