Fix dangling pointer in simpleperf_manager.cc
We were not assigning ProcessManager::GetPackageNameFromAppName(), nor
(CurrentProcess::dir() + kSimpleperfExecutable) to a variable and were
calling .c_str() directly on them instead. These strings were going out
of scope, resulting in dangling pointers. Consequently, we were passing
unwanted values when invoking simpleplerf (e.g. a path instead of name
to --app option, or an app name instead of simpleperf path).

The fix is simply assigining the method calls to a string variable and
use this object when calling .c_str().

Test: Manual
Bug: 37294023
Change-Id: I69074c3cdb6f89727f458208871b2ddc0724956f