66508267: Version comparison is not accurate for versions with dashes
The architecture components were recently released with the version
name "1.0.0-alpha9-1".

The "-1" causes problems for our GradleVersion class, which does not
match dashes in the version number.  However, the semantics for
GradleVersion is not correct; Gradle/Maven does alphabetical sorting
after the x.y.z- version strings; we should do the same.

Test: Unit test included
Bug: 66508267
Change-Id: I39bc3ff42a239422875c716ea9405bfb19d8eb4e