Small act-d tweaks.
Increase retries from 3 -> 5. I noticed one of the recently build
failures was due to getting a 503 from act-d. More retries is a low
effort solution. A longer-term solution will be to migrate act-d to
a more dedicated hosting.

Also add a success message after act-d information has been uploaded,
so that I can search it in the logs to make sure it actually happened.

Test: presubmit
Change-Id: I9985fae5e3a14b8c989f740bba9b6f87b5b9d284