Add support for argument to parameter mapping in Kotlin
This CL adds a new Kotlin-specific library, lint-kotlin, containing
utility code intended for lint, which can access the Kotlin PSI.
This is not exposed as part of the lint, but lets lint implement new
services, such as a full call argument mapping. This is used by
the annotation checker to correctly correlate arguments to a call
with the right parameters. In Java this is a trivial map, but in Kotlin
thanks to extension methods, default parameters and named parameters
the argument order can vary wildly from the method parameter order.

Also adds README for the new lint class loading architecture,
and fix firebase lint rule issue suppression (bug 67986477).

Test: Unit test included
Change-Id: I95df81f8ac6776fe149d5730e2e479378aba1f7f