Standalone Gradle plugin for Lint
This CL adds a Gradle plugin for running lint on projects that do not
apply one of the Android Gradle plugins (library, application, instant, etc.)
In fact if you do attempt to apply both, the plugin will fail the build
with an explanatory message that lint is built in for the Android plugins.

The purpose of this is to let teams use lint on projects that are unrelated
to Android, such as a pure Kotlin project.

Options can be configured in the same way as for Android projects -- using a
lintOptions { } block (but unlike with the Android plugin, this is specified
at the root level instead of inside an android { } block.

Similarly, they can depend on lint custom checks with a lintChecks
configuration.

Test: Integration tests included

Change-Id: I33af970fcff248cda1f8187eeb7ec5e4605a3d75