Update proto.bzl for bazel 0.8.0
In bazel version 0.8.0, += now mutates the existing list as opposed to
creating a new one. This means lists and dicts passed from the BUILD
files can't be operated directly on by += because they come from a
different context and frozen. They need to be copied to an ordinary
list/dict first.

java_proto_library is the only violator.

Test: n/a
Change-Id: Id0d61fcbd2d52000a1b6020e3412bf250d204a28