Fix ClassNotFound error in PathUtils.deleteIfExists().
Sereral tmp directories have a shutdown hook that allow them to be
deleted when the JVM (Gradle daemon) exits.

The shutdown hook currently uses an inner class. If this inner class is
never loaded during a build, trying to load it when the Gradle daemon
has been shut down could result in a ClassNotFoundException.

We fix it by writing an equivalent implementation for
PathUtils.deleteIfExists() without relying on an inner class.

Also make PathUtils.addRemovePathHook() public as it seems useful, and
let several places use it.

Bug: 69380617
Test: PathUtilsTest + manually verified that it fixes the sample project
    linked from the bug above

Change-Id: I15ccf413599d50fa199d1cabb7332a3827a825a7