Support directories as inputs in FileCache.
FileCache used to support only regular files as inputs. In commit
8ae03e648dd8661158941f8449f8dc26ce2990cf, we began to support
directories as inputs as well. However, that commit was incomplete,
resulting in bug 70211407.

To support directories, we need to clearly define what should be
part of the hash of a directory, including the file contents, the file
metadata (file names, modified time, etc.), or the directory structure.

In this commit, we compute a directory's hash from the following:
 - The relative paths, in lexical order, of all the regular files and
   directories at all depths below the given directory (symbolic links
   are not followed)
 - The contents of all the regular files in the sorted list above

Since computing a directory's hash can be expensive, we need to watch
out for potential performance issues introduced by this change.

Bug: 70211407
Test: Updated tests are included
Change-Id: I7a50247c42fc4594a68b8d34c1753b6dad1a22ba