Revamp aapt namespaces support
- Introduce ResourceNamespace class to represent the concept of an aapt
namespace, which can belong to a package name or be a special case like
the res-auto namespace (which is also used for projects that don't take
advantage of namespaces). This makes it clear in the type system when
we're dealing with resolved namespaces and when with short forms taken
from the XML.
- Use ResourceNamespace as the first dimension of the resource
repositories. Using RES_AUTO for non-namespaced projects means we don't
have to worry about nulls not allowed as collections keys.
- Extend ResourceReference to keep track of the namespace and type of
the resource it's referencing. This way we have an object to pass around the
coordinates of a resource. Also, ResourceValue already extends
ResourceReference, which is nice.
- Add ResourceUrl#resolve(...) method which can be used to turn one
namespace representation in the other, given enough information about
the context. This makes it clear that a ResourceUrl cannot be used as-is
when namespaces are involved.
- Add ResourceValue#getValueReference() method, which handles namespaces
and should be used instead of manually parsing the textual value as a
ResourceUrl. Knowledge of defined namespaces is passed to the
ResourceValue from the creating code, i.e. ResourceItem or
PsiResourceItem.

Test: existing tests, AppResourceRepositoryTest#testNamespaces
Change-Id: Ide7a1bd841b926fe357ea8416b0f18bef6fedd35