Begin breaking apart the GradleTestProject testDir weirdness.
GradleTestProject has a field called testDir, which refers to the
directory to which the project is copied before a test (I plan to rename
it to projectDir in a future CL, to better represent what it's for).

This field begins life unset, and is only set inside of the apply()
method, which is called by JUnit before a test is run. This creates a
temporal dependency where a swathe of methods on GradleTestProject
simply don't work before testDir is set.

In an ideal world, this temporal dependency would not exist. Sadly,
GradleTestProject is a complicated class and I don't know exactly how
much work would be required to totally eradicate the weirdness. I can,
however, avoid it for the BenchmarkTest suite by setting it explicitly
before a benchmark is run.

This has the side benefit of moving BenchmarkTest one step closer to not
depending on JUnit.

Test: these are tests.
Change-Id: I79f3f83984a5249a41561877c97d70b4b28f6724