Fix SigningTest on Windows
Removed the '\n' from the "search" regex in failWithMissingKeyPassword
because, while TestFileUtils.searchAndReplace() automatically replaces
'\n' with '\r\n', TestFilUtils.appendToFile() does not perform the same
substitution. This was leading to a failure, even though the regex was
identical to the string that was added to the file in the first place.

Test: Test-only modification
Change-Id: I17409469feb7e5e2425cc012010684a69918424e