rework of the BuildableArtifact
simplified API to not require pre-declare output files. Output files
are now specified in the config action using the output provider.

changed API to disambiguate final inputs (as the final result
of actifact transforms on a type) versus inputs that are appended or
replaced in the chain of build artifact transforms.

removed the notion of first artifact, all artifacts are now treated
the same and the first version is defined by being the first to
register for a particular type. As of today, the plugin gets
a chance to register these first versions first but we could
envision a future public API for end users to allow being first.

Test: updated existing, added new ones.
Bug: 73602296

Change-Id: I526da9ee6da45fe417c05ac51ca9de57bd1fcb63