Move computation of feature names to the base module.
This allows to ensure that they are all unique. They are published
back to the feature via the the old FEATURE_IDS_DECLARATION which is
renamed FEATURE_SET_METADATA as it contains metadata for the whole
set of features.

Also rename the class writing/reading this from FeatureSplitPackageIds to
FeatureSetMetadata.

Change the way tasks that need the feature name or the per-feature res ID
(renamed ResOffset). The API now creates a supplier that returns the exact
value for the given module. It also handle making the task depend on the
artifact published by the base module.
This has 2 advantages:
- remove the need for each task to have the full file as an input which
  would trigger some task execution even if the file only changed in parts
  that don't matter for the task.
- remove the duplcated parsing code from all the task

Finally, tweak VariantType to add isFeatureSplit. In general it's better to
test variant type instances for positive attribute rather than for false values.
For example: isBaseModule == true means that this is a base module. However,
isBaseModule == false does not mean that it's a feature split/optional module,
because tests, and library, etc... also return false.

Test: WIP
Bug: 74177590
Change-Id: I37889dfe7776a740eb8495c490a19f36d5d29556