Add a basic relocatability test.
The Gradle build cache allows you to cache the outputs of build tasks
locally and even across machines. To do this, Gradle relies on plugin
writers to craft a set of stable inputs to be used as a cache key. The
inputs to the cache key typically correspond to the task inputs, but
things like incorrect use of PathSensitivity can lead to builds that
don't correctly cache across the same build on different machines.

This test aims to be a first step at actually testing for relocatability
of builds. The general principle is that you should be able to build the
same project in two different locations, with the second build being
served entirely from cache.

That is not the state we're currently in, so the test maintains a list
of KNOWN_PROBLEMS that we should work through over time.

Test: this is one
Bug: n/a
Change-Id: If8656c7b66046a33c33ce9f104caeeee62b56ef5