Revert "Jetifier support: Transform libraries to use Androidx."
This reverts commit 7a1259e0a961ecd685de9248ebfe8923d70bb296.

Reason for revert: The jetifier-core library is not yet available on
remote repositories (only in prebuilts). AGP cannot use the library
until it is available on jcenter() or google(), or until we repackage
the library as part of AGP.

Only this commit needs to be reverted. The other commits of the same
topic can stay without any harm.

Change-Id: I9b160ff8e9e6daded144be628c1eb21613ce82b4