Use Gradle module name as feature name.
Renaming them on the fly to handle conflicts makes the
module name too magic and hidden from developers. Since
the module name is needed to handle dynamic request of
on-demand module, the name must be consistent always.

This changes removes the de-duplication and instead forces
all feature modules to have a unique leaf name in their
gradle path.

Bug: N/A
Test: existing tests.
Change-Id: I13e069aa4489543efbcaf6ff170c75c7be1a74a0