Send energy events even if profiler attaches late
Subsequent wake lock, job and location events will stop checking if the
event ID already exists so if profiler attaches after the first event we
will still send subsequent events.

For alarm events however, we depend on the first set-alarm event to
determine if they are wake-up alarms and therefore we must keep the
existence check.

Change-Id: Ica099783a851e31ae0f986451e79e0160fcf90a2
Fixes: 79431656
Test: perf-test/EnergyTest