Factor out JSON generation immutable state
Bug: None
Test: No behavior change expected

This is a simple refactoring I've intended to do for a
little while. It pulls all of the immutable JSON generation
state into a separate container class. It simplifies the
code and reduces the risk of passing so many parameters
of the same type next to eachother (soFolder vs. objFolder
for example).

Change-Id: I8a4713fb272fd0000bd442302f76193dc4ef03cd