Fix JniTest flakiness that Jvmti data is separated in multiple calls.
The test waited for 10 ref events in perfd data, but when a call
does not have 10 events yet, it uses the end timestamp as the next
call's start timestamp, then assert the two repeat does not have
overlap events. But the test logic is too complicated and it is
flaky when overlap events happen. Use a simple test logic that each
time of GetJvmtiData to assert the ref count, the result does not
depend on previous GetJvmtiData response.

Bug: 80501379
Test: JniTest
Change-Id: Iadd91c8ff5397d5917cff1bf411be9c8f497396a