Remove unsafe build cache locking for older AGP versions.
Since AGP 3.1.x, we've created one private cache directory per AGP
version under the shared cache directory, as shown below.
./android/build-cache
|-- <cache entries created by 3.0.x and earlier>
|-- 3.1.x
    |-- <cache entries created by 3.1.x>
|-- 3.2.0-dev
|-- 3.3.0-dev

Since the private cache directories of AGP 3.1.x and later and the cache
entries of AGP 3.0.x and earlier are in separate locations, concurrent
access is almost always safe.

The *only* instance where it is not safe is when a plugin with version
3.0.x or earlier is deleting the shared cache directory while a plugin
with version 3.1.x or later is using a private cache directory inside
the shared cache directory.

Previously, we attempted to work around the issue by locking the shared
cache directory for the entire build and unlocking it at the end of the
build, but ran into an issue where the unlocking would fail because the
thread that does the unlocking is not the thread that did the locking
(unlocking happens in one of Gradle's buildFinished() event handlers
which could run on a different thread than the thread that executes the
tasks---see https://issuetracker.google.com/80464216).

This CL removes the workaround because: (1) That scenario should be
rare; (2) It will never happen once the user has updated all their AGP
versions to 3.1.x+.

Fixes: 80464216
Test: Existing BuildCacheUtilsTest + Manually verified on the repo
      project

Change-Id: I0f9623b8416b5016d8f226015760b8748bb3d8bb