Fix trace with unicode symbols not being parsed
We've recently started supporting unicode characters in trace files parsed
by VmTraceParser. However, we forgot to handle the special case where a
unicode symbol is read from the BufferedReader and takes 2 bytes to be
represented (BufferedReader reads characters in the range 0x00 to 0xFFFF).

The issue was we were reading 2 bytes but only 1 char, so we're wrongly
incrementing our offset by 1 when it should be incremented by 2. As a
result, we were failing to verify the magic number later on and the trace
was not being parsed.

Incrementing the offset by the number of bytes (instead of characters) read
fixes this bug.

Bug: 110564368
Test: VmTraceParserTest#testCharWith2Bytes
Change-Id: I68f7a1a8a0d2eec41593f7a954cee0066ea3e84e