Fix remaining gradle foundry failures by installing 32 bit libs on docker image.
The last of the gradle integration tests that are failing are all failing
because the prebuilt linkers in the SDK have dependencies on 32 bit libs. This
CL installs those libs (i.e. libdl.so.2, libz.so.1, libm.so.6, libc.so.6) in
our docker image and removes no_test_foundry. These tests currently pass on
foundry with this docker image.
See go/upsalite-prod/ef42d2f6-cfd5-449f-a502-3ed51a55a943 for failing tests.

Bug: 110432155
Test: bazel test --config=remote
        //tools/base/build-system/integration-test/application:automatic-tests
	//tools/base/build-system/integration-test/application/src/test/java/com/android/build/gradle/integration/nativebuild:RsEnabledAnnotationTest
	//tools/base/build-system/integration-test/application/src/test/java/com/android/build/gradle/integration/nativebuild:RsSupportModeTest

Change-Id: I707e8933795aa5c2e7f4ebb1aca3a5d342011813