Revert the revert of upgrading RBE stack.
This reverts commit eb80da604ad9f5bef17662da6d2b3ef264777d18.

Reason for revert: This is re-landing a migration from our existing
RBE pool to a new RBE pool at the request of the Foundry team. This
is a revert of a revert, because we tried it once, saw unexpected
performance and then reverted to go back to the old pool. We believe
the performance issues have been fixed, and so we are trying again.

Old pool usage dashboard:
https://viceroy.corp.google.com/alphasource_foundry/customer_scheduler?project_num=764168210157

New pool usage dashboard:
https://viceroy.corp.google.com/alphasource_foundry/customer_scheduler?project_num=854894812152

Change-Id: Ie50680eda26b34f7427b2be524194ee300be6600