Add a two-argument constructor to Variant
This simplifies construction from Blaze and unit tests, where the
Variant is just a container for a list of Artifacts. In such code,
the Artifacts are already in a flat list and will need to be attached
to different properties based on their artifact name. This created
a bunch of boilerplate. This change moves the boilerplate to a common
location, in a new constructor.

Test: VariantTest
Change-Id: Ia762b45cac6614b4b99e2955f596ce04d6e45ba2