Incap: Create ProcessAnnotationTask only when required.
In commit 3a75752, we had a small optimization where we bypassed
ProcessAnnotationTask's configuration if it is not needed. However,
because some lateinit properties may still be accessed at execution
time, the build may fail complaining that the properties have not been
initialized.

This issue was not caught in IncrementalJavaCompileWithAPsTest because
the test scenarios missed a corner case.

This CL will determine whether the task is needed up front, and create
the task only if it is needed. If the task is executed, it will always
have been configured fully.

Also update IncrementalJavaCompileWithAPsTest so that it covers the
previously broken scenario.

Test: Updated IncrementalJavaCompileWithAPsTest
Fixes: 116170968
Change-Id: I86c884b4514f22337035e58e5ad8efb62804938d