Auto namespacing: Use correct package for attrs in R class
When we bulid the SymbolTable for each library, we do not know yet if
the children under the declare styleables are refering to local
attributes or to remote ones. That's why they do not have namespaces
(other than "android:"), and when we want to write the R class we need
to make sure we use the proper namespaces for each of them.

For example:
<declare-styleable name="ds1">
    <attr name="android:color"/>
    <attr name="remote_attr"/>
    <attr name="local_attr"/>
</declare-styleable>

will result in the R class containing:
int[] styleable ds1 = { 0, 0, 0};
int styleable ds1_android_color = 0;
int styleable ds1_com_remote_package_remote_attr = 1;
int styleable ds1_local_attr = 2;

Test: added new and existing ones should pass (especially
AutoNamespaceTest and AutoNamespaceDependenciesTest)
Fixes: 113319396

Change-Id: I26597bb0e1cf42b3d6dada18460ca3d9c5232afc