Update DeviceConfig, which pulls info from a physical device
DeviceConfig is an app that runs on a physical device and
writes an XML file describing characteristics of the device.
These characteristics are needed to emulate that device.
Parameters include screen size and pixel density, memory
size, camera, sensors (compass, accelerometer, etc.), hard
keyboard, etc.

The resulting XML file is typically pulled from the device
using ADB. It is used to create a <d:device> entry in a file
such as sdklib/devices/nexus.xml.

Bug: 117556079
Test: Verified with Pixel 3, 3XL, and emulator
Change-Id: I5670a7e780433288eb7995ade5bf658c601e68a7