IRR: Add delta push capability
DeltaPush aims at reducing the volume of data sent to the device
while pushing a .apk file. Since apks are zip archives, deltaPush
leverages the Central Directory information to determine what has
changed between two apks and only transmit a minimal "patch" which
is used on the device to re-construct the apk.

Bug: 119576989
Test: Unit test + manual tests
Change-Id: I395425a258fdcd52d0b5a1c519120845ed515223