Remove deprecated NdkCompile codepaths
NdkCompile was an older codepath used to support C++ in AndroidStudio.
It has long been deprecated and the user has received a warning about
this. In 3.3 we introduced a 'lease' flag that the user would have to
periodically reset to continue using. It's time to get rid of this
old code.

Bug: NA
Test: Code removal
Change-Id: I2fa0c88e25d3a2f37c7fb9e572b298a49712ce8e