Delay registering an Http event until the request starts
Previously, an Http event is registered on pre-connect (when
the HttpConnectionTracker is instantiated), and additional data
including the request method and fields are sent at a slightly
later time (when the request starts). In practice, these two stages
are close together.

This change moves the registration of the Http event to the later
stage to simplify how the data can be sent via the new data pipeline.
It means that http connections that are created but not actually used
will not show up in the Network Profilers anymore. That should not
have any major implications because there will be no data to show
for the connection anyway.

Also updated some comments and removed obsolete TODOs to be be
consistent with features that have already been implemented.

Bug: 117826867
Test: Existing test already expects the updated fields to be
available together.

Change-Id: I3dbbb9e2931968478f7d96301beab2bbf2372ceb