Enable choice of default Variant in the DSL
Allow the build author to pick their build types and product flavors
they would like to be selected in Android Studio when opening the
project for the first time.

For projects that are shared widely, this removes one manual step from
the setup.

For example, the build author can make their custom build type the
default using the following DSL:

android {
   buildTypes {
       dev { isDefault = true }
   }
}

For cases when the build author does not specify, or only partially
specifies the default variant dimensions, it is also adjusted to pick
'debug' as the build type if is is present, otherwise the first
alphabetically, and the first alphabetically of each flavor dimension.

Given build types 'alpha', 'debug' and 'release', and flavors 'a' and
'b', previously Android Studio would default to aAlpha, but now it will
default to aDebug.

If the user selects multiple build types or product flavors of the same
dimension to be the default the first alphabetically is used, and a sync
warning is output.

Bug: 36988145
Test: Added DefaultVariantTest (functional test)
Test: Updated CompoundSyncModelBuilderTest (integration test)
Test: Updated relevant unit tests.
Change-Id: I8d36140f710ee4426af7bef40ec5ebdc104ad0d7