Support CLASSLOADER isolated worker through Facade
The WorkerExecutorFacade supported only NONE isolation mode so far
forcing tasks like Jacoco to use the Worker API directly when
needing to spawn the worker in a separate classloder.

Adding this feature allow these tasks to still use the facade
and provide the desired isolation mode (only NONE and CLASSLOADER
are supported so far). This allow for proper attribution of the
worker to the spawning task.

Test: Existing.
Bug: 126731735
Change-Id: I4e55d53c16030a68ae74a7f033c94354c64c3446