Improve performance of incremental support for zip inputs.
ZFile is terribly inefficient in the way it reads Zip file
because it tries to be as robust as possible when reading
zip files.
In order to support incremental changes to entries inside
zip archives we need to parse zip files as quickly as possible.

This provides a new implementation that only parses a Zip file's
CDR and allows writing it back into a new zip without the actual
entries (to improve the zip-cache).

This also attempts to re-use in some places the parsing of
Zip files between pre-work up-to-date computation and
post-work incremental state saving.
In previous implementation Zips were read 3 times:
- to compute what file changes
- to read the file for task processing
- to store the new state of each file.

Change-Id: I4db39b2f74f94d0474e1bf0ae35f5ff4552eacbf
Bug: 128031248
Test: Existing + more TBD.