Remove false error in Event Profiler's logcat
When ActivityInitialization registers Activity lifecycle
callbacks, it can do so only when the Activity has
launched. Currently when it retries doing so, it writes
an error to logcat if the Activity isn't available yet.

Momentary unavailability is expected and should not be
an error in logcat. The error may confuse the user.

Bug: 131352670
Test: manual
Change-Id: I6f0502b0c798966c4cc4f01047e0bd487a03bde3