Have Daemon be responsible to generate the CPU trace id again
This change partially undo 45131d196edaf8b860a0b8208e55d3ab628d5be2
where the app agent became responsible for generating the cpu trace id.
The intention of the change was that in the new pipeline, the commands
need to generate and cache the trace id in the app agent, but upon
further review that is no necessary.

Bug: 119261457
Test: Existing
Change-Id: I48880ee4ac128e0102c7dbef7c726b1ff4cdbd66