[aapt Kotlin] Added parsing Floating Point types
The focus of this diff is on parsing Floating point, Dimension, Fraction
and Color type resources. Because the logic of parsing floating and
fixed point types is particularly agregious, I wanted the logic for this
part to be in its own CL.

The process of parsing a Floating point and Fixed point is done as
follows:

1. Identify a suffix for fixed point (it is a floating point if no
suffix is found.)

2. Verify suffix does not interfere with floating point parsing. i.e.
0x20.dp. Is it 0x20.d with suffix p? Or 0x20. with suffix dp. On
ambiguity we should fail. (consistent with C++ parsing)

3. Ensure that all string values that can be interpretted as floats in
C++ can be interpretted as floats in Kotlin. e.g. 0x23 can be
interpretted by C++ but not in Kotlin.

4. parse the float.

5. reformat as Fixed Point if necessary.

Bug: 132800341
Test: Tests added for ResourceTypes (parsing Floating Point values), and
ResourceUtils (parsing Color values)

Change-Id: Ifb9cff6f00d58732dc1cbaa6f7897c0a6a000375