Jetifier: Skip libraries that already support AndroidX.
Jetifier 1.0.0-beta06 adds a new feature to bypass converting libraries
that already support AndroidX:
    https://android.googlesource.com/platform/frameworks/support/+/f9fd8a5c069d49fe023af9b04b54457260a32417

With this CL, the AGP now uses this new feature by default. However, as
a safety measure in case the new feature has a bug, we provide a
property called `android.jetifier.skipIfPossible` (enabled by default),
which the users can disable to revert to the previous behavior before
this feature was introduced.

Bug: 119135578
Test: The feature is already tested in Jetifier
Change-Id: I3c6dbc840a573a78ad414b37ed3e6fcb42dd78b6