Use environment variables from BuildSettings.json
BuildSettings.json is a peer to CMakeSettings.json that allows users to set environment variables for the Ninja build. The settings will be read with each build and is stored in a new field in CxxAbiModel. The model is accessed in 2 parts:
1) During CreateCxxAbiModel, which reads BuildSettings.json and gets a
BuildSettingsModel.
2) During ExternalNativeBuildTask, which uses BuildSettingsModel to
apply the environment variables to the Ninja process.
The model is at the ABI level because this is where macros are expanded.
To keep CL sizes small, macro expansion will come in a separate CL.

Bug: 137211488
Test: Unit/integration tested

Change-Id: Ic8c3eec08a83f851315a1539a6a080a76d0c02c8