Don't mutate EmulatorConsole port number.
Currently EmulatorConsole#mPort is a mutable int because
EmulatorConsole's termination logic sets it to -1 to indicate
the console has already been closed.  This port
number should not be mutable because the EmulatorConsole's
console cache relies on it to correctly add/remove cached
instances.

The port number state is now immutable and the console closing
logic that used to rely on the mutability of the port number
has been simplified such that it's no longer required.

Fixes: 138780921
Test: EmulatorConsoleIntegrationTest added. Tested for flakiness with
1000 runs with 0 failures.

Change-Id: I075f4adf5d5f6d6763534ba6c0785594e0502354