Make Studio be responsible for calling StopProfilingApp before StopMonitoringApp
This change helps to align the legacy and new transport pipeline. In the new pipeline,
we need to call stop tracing explicitly before ending a profiling session. StopMonitoringApp
will be going away so separating the two logic is cleaner anyway. Added an extra field to
both CpuProfilingAppStopRequest (legacy) and StopTraceCommand to indicate whether
the daemon needs to wait for the trace response.

Bug: 135553534
Bug: 133324501
Test: Studio-side test
Change-Id: Idc0c71c76404e255d7381809eb70bd60f7397eab