Skip process-annotations task when no processors exist.
In a sub-project where no annotation processors exist, the
process-annotations task would produce the following warning:

    "Annotation processing without compilation requested but no
    processors were found."

That warning is not really needed, and can turn into an unwanted error
if the user sends "-Werror" to javac to treat warnings as errors.

To address that, this CL will skip the process-annotations task in that
case.

Fixes: 140602661
Test: Updated IncrementalJavaCompileWithAPsTest
Change-Id: Ieedb05bf4efd1e1c5d195ecd3f2c358282a1e2ec