Add missing task property annotations.
Gradle 6.0 will show task property warnings to users more visibly (it
will show that the build contains warnings, and if the user enables a
flag, it will list all the warnings to them).

Therefore, it is important to fix all task warnings now (and also for
AGP 3.6 in case AGP 3.6 users update their Gradle version to 6.0).

This CL:
  - adds missing annotations to task properties that are "easy"
  - assigns @Internal annotation to "difficult" ones to revisit them
    later (those locations are searchable by the bug ID: 135900510)

Bug: 135900510
Test: ValidateTaskPropertiesTest
Change-Id: I94437f18eeda0034122cc2c75ab6b7c71958aa91