Dup file descriptor when agent connects to daemon
When using layout inspector, if the user closes the
project and reopen it, the layout inspector sometimes
stop working. The CL fixes the bug.

When daemon and agent establish the communication channel,
the daemon sends the agent a file descriptor, and the
agent would build a GRPC channel on top of it.

When a Studio project is closed, the daemon is killed. But
the agent (with the app) keeps running. If the project
is reopened, the daemon will sends the agent a fd again.
If it happens to be the same as the one used before,
the channel cannot be established successfully.

Commit bfa6eff10c9e87a6c6812c35930ef3e538459e74 attmpted
to fix the issue in Year 2017 but the implementation may
not work 100% of time. This CL is another solution.

Bug: 143543252
Test: manual; close-reopen project multiple times and
      layout inspector consistently works.
      Verified profilers still work for pre-O and O+.

Change-Id: Idb8b1554ac83affcbf1f21feca378075f0263638