non-main artifact dependencies in GradleModelMocker
In the mocker, support dependencies with slightly less restricted
configurations.  We support the compile and implementation scope for
the main artifact and for test and androidTest artifacts; in order to
do so, add a number of mocks and fields to allow lookup of the
artifacts from the variant, and dependencies from the artifacts.

Bug: 124415929
Test: new unit test
Change-Id: I37a6988bc9f23e19f0db763f7c686f139e7417b9