Fix extracting literals from annotations in compiled JARs
The current implementation of extracting literal passed as Annotation
parameters do not work when the Annotations come from a compiled JAR.

This CL splits the literal extraction logic and uses the appropriate logic
when dealing with a compiled class instead of a source.

The current literal extraction logic uses UAnnotation class to extract
the literal. Kotlin and Java sources are serviced by KotlinUAnnotation
and JavaUAnnotation respectively. However, compiled class are serviced
by Cls* classes, and ClsUAnnotation does not exist. This CL creates a
new sealed class AnnotationValuesExtractor which has `Source` and
`Compiled` subclasses containing logic to extract literals from sources
and compiled classes respectively. `Compiled` AnnotationValuesExtractor
uses UAnnotations's lower level PsiAnnotation (`javaPsi`) field to
extract the constant. This approach does not work for Kotlin, so two
implementations are needed.

Fixes: 142135283
Test: testPermissionAnnotationsInCompiledJars in PermissionDetectorTest
to test if @RequiresPermission(...) is resolved from within a compiled
JAR.
Test: testTargetApiInCustomJar in ApiDetectorTest to test if
@TargetApi(...) is resolved from within a compiled JAR.
Test: testRangeAnnotationsInCompiledJars in RangeDetectorTest to test if
@IntRange and @FloatRange are resolved from within a compiled JAR.

Change-Id: I9719fe2bccfec6cb78dd243a4e4153819bd35fb5