[Wizard Templates] Fix resource loading on release build
When running on "proper" release build
the resources are packed inside
"lib jar files" and loaded by a special
"plugin classloader". To find the resources
we need to supply a class that was loaded
by the right class loader.
When loaded this way, the resources need
to have a full path (ie start with "/").
This CL improves on Change-Id Ie35f51e420dce07b51f55ca79026e77eb21b370a

Bug: 135397531
Test: TemplateTests + manual from a release build
Change-Id: Ib17d3d3083b868bc99891be98d0731ab86658cd5