Allow UNKNOWN_ARCH for instrumentation APK dump.
When we dump an instrumentation package, we will never
find a process running in that package name. Therefore, it
will always have ARCH = UNKNOWN. Fortunately, we also
do dump for the target package id which should be running
the app with instrumentation. The problem right now is that
the dump will have an actual ARCH "mixed" with UNKNOWN ARCH.

This CL lets ApplicationDumper to run as long as there is
exactly one known ARCH.

Bug: 144856589
Test: ApplicationDumperTest.testProcessIds() and testMixedArch()
Change-Id: I0999bc70143ef674818d7e2bcdb245de47d0a6c8