Clean up class hierarchy for ProductFlavor/BuildType.
CoreProductFlavor and CoreBuildType are now deprecated.
They cannot be fully removed because they are leaked into our
public API from AndroidConfig.

DefaultProductFlavor and DefaultBuildType are also
deprecated, made abstract and renamed Abstract*.
A lot of their usage was cleaned up in favor or more specific
types.
They are still used in a few places that need to be cleaned up
before we can hope to remove them.

Bug: N/A (clean-up)
Test: existing tests
Change-Id: Ieab8377ae191389d37d4fac2b2a2c832e84cbecb