Switch from model.builder.SigningConfig to internal.dsl.SigningConfig
In order to address b/134858541, we want to add addiontal information
(e.g. whether v1/v2 is configured by users) to dsl.SigningConfig but not
to builder.SigningConfig. However, we currently use builder.SigningConfig
when writting signingConfig to json which prevents us from writting
additional information added on dsl.SigningConfig. So this commit is
an internal clean up to switch from builder.SigningConfig to
dsl.SigningConfig.

Bug: na
Test: existing
Change-Id: I2b095ed197542e6e46218d1342fa6532a86ab9a7