Update to latest UAST (1.3.70)
This CL updates lint to the latest version of UAST from the 1.3.70
branch in the Kotlin repository. With this we pick up some very
important bug fixes, such as the one for KT-32031; prior to this all
suspend methods were completely missing from the AST. There's now
a new passing test for this scenario.

This also moves UAST over to platform 193, which required some big
changes in the setup code. It's no longer possible to subclass the
application environment, so instead, lint now has a UastEnvironment
class modeled after KotlinEnvironment in the Kotlin compiler, which
attempts to delegate as much as possible over to KotlinEnvironment.

Finally, the new PSI environment also required the Kotlin type
resolution to happen up front before anyone attempts to call into any
of the application services; this was for example done by some
resource detectors (like DeprecationDetector) to do inheritance
lookups on classes referenced in XML files. This required splitting up
the UAST visitors into two halves: first the setup part, run prior to
any XML processing, and then the actual visitor part, which still
needs to happen after XML processing.

There are some changes to the shape of the UAST; for example, some
lambda expressions in Java went from this:
  ULambdaExpression
    UCallExpression
      UIdentifier
to this:
  ULambdaExpression
    UBlockExpression
      UReturnExpression
        UCallExpression
          UIdentifier

This required adjusting some lint checks.

Also updated the code to rewrite a number of now deprecated
constructs, such as UastContext and UAnnotated#annotations, to the
recommended replacements.

Test: Many unit tests added and updated
Fixes: 135168818
Fixes: 143972324
Fixes: 141622949
Fixes: 140328642
Fixes: 142952555
Change-Id: I0ba284f52e62015af95aff8c5e83539b5a888e81