Fix local-module custom Lint checks on IDE side
The location of lint.jar files in the Gradle build output directory
has changed again. This time the hardcoded paths in Lint were updated
to match (see change I0ed76e7d1), but unfortunately an older path was
also removed, which broke backward compatibility with previous Gradle versions.

The root cause is b/66166521, but for now this change simply
updates the hardcoded paths.

Bug: 146403481
Test: existing and manual
Change-Id: Ief5b79b70bc1beeb00831bb14b4e357743e3fafe