Add avd rule and test for emulators on RBE
The avd rule allows users to launch Android emulators for tests both
locally and on RBE.
A sample showing how the rule is used is in bazel/avd/sample.
The rule requires the emulator binary, which is checked into prebuilts
on Linux as well as the system images for SDK level 29, which are
downloaded from dl.google.com when needed by Bazel.
The Docker image update installs dependencies needed by the
emulator and increases size of the image from 2.07GB to 2.37GB.

Bug: 136178744
Test: EmulatorTest.java
Change-Id: I76262ef2f2890784e184511193a47c20134175ed