Add Prefab publishing options to AARs.
One thing to note here is that Prefab libraries can't be shared
across modules within the same project. The reason for this is that we
need the Prefab package to be generated from the dependency to
configure the build for the dependent. Generating the Prefab package
is something that requires that the module has built, since it
includes the built libraries. We can't usefully use fake stub
libraries for this because the dependent might configure itself based
on the contents of the built library. As such, we need to build the
dependent module before we can configure the dependency, which is not
compatible with Gradle sync.

Test: Added new tests
Bug: None
Change-Id: Ie140e17edebf157f184ecaa4587172008b15b37b