Speed up ResourceTable
Fix algorithmic growth rate for the Resource Table Class. The typical
use case for the Kotlin Resource Compiler is a large amount of resources
in a single file rather than many small files. As such, we are moving to
optimize the resource table for a large amount of entries, in order to
improve performance.

Test: integration tests still pass.
Bug: 132800341
Change-Id: Iced74c8681bd591214f26b4ba29b9122020ac2af