[iml_to_build] Fix runtime library inclusion
With this change we include library dependencies in the transitive
closure computation. This needs to be done because the iml_runtime
rule doesn't collect dependencies transitively (on purpose, as some
of them are there in order to break cycles), so we need to bubble
them up.

Previously only modules were considered, resulting in many targets
missing a few runtime libraries from their classpaths. As it turns
out, we mostly don't trigger any behaviors from these runtime
libraries, so their absence is not noticed.

With the recent migration of the IDE's tips-of-the-day feature to
a Maven library, included only as a runtime dependency, we ended up
with many tests breaking during parsing of the problematic inclusion
in the plugin.xml.

Bug: 148420958
Test: added
Change-Id: I0abc2fbd7d377cdeef352f81e4fbd4cf0e31dfe9