Adding unit test support for Agent
Also fixing a couple small bugs in layout inspector:
- Previously we were checking to see if DetectRootViewChange should be
stopped before sleeping and not after. This meant that on disconnect
we'd mark it as stopped, sleep, then the root would refresh, and then
DetectRootViewChange would wake up and detect that and start up the
listener again.
- Previously we weren't setting the pid on events. This didn't usually
cause any observable problem because we weren't correctly filtering on
pid from the studio side, but when switching between two devices
(combined with other problems on the studio side) it could cause us to
receive events from the wrong device.

Fixes: 147491720
Test: Added framework and initial test for LayoutInspectorService. More
tests coming in followup cl.

Change-Id: I4301561012ea9cb649847fcc8f5c02c662408e5a