New Manifest parsing mechanism for variant values
With the new variant Api, we need all the values that
can come fro the manifest to be using Provider<> for
their lazy mechanism.

On top of this, to prepare for the day where the sources
are transformable via BuildableArtifact (which means
representing sources as Provider<FileSystemLocation>),
it's important to keep possible task dependency
throughout.

This change introduces a parallel manifest parser that
uses Provider<RegularFile> to create a Provider<ManifestData>
which then can be mixed with the DSL values.

Once it is fully implemented (only versionCode/Name,
and test attributes are supported right now), the old
manifest parser will be removed.

This test also removes some integration test that
validates version names since this is now handled
by unit tests

Bug: N/A
Test: existing + new unit tests
Change-Id: Iaf5f47c3c94a58382396edfa95ad6cc3e502ed37