Fix regression with instant-exec.
when using instant exec, with task graph serialization, the
ApkFormat is incorrectly deserialized into an object which is
not part of the ApkFormat.values().

The equals() would then fail as the object instances would
be different. Using the semantic name() comparison is guaranteed
to be correct.

This created a regression in Base100_Java7_vfsAndInstantExec_codeImplChange
benchmark as we started using apkzlib instead of zipflinger when
re-running tasks from serialized task graph.

Bug: N/A
Test: Existing
Change-Id: Ie6996b9d08a5484677e7f7a857d267c2ca2527fe