Zipflinger: Don't rewrite CD on untouched archives
Problem: Zipflinger operates under the assumption that an archive
will always be modified. It always rewrite the CD at the end of
an archive. This is an issue now that it is used to parse and
extract file in Apply Changes 2 to generate overlays since it
results in signature removal.

Solution: Only re-write the CD when the archive is being created or
when the archive has been modified.

Fixes: 152807391
Test: Unit tests
Change-Id: I22ece3476e9b35bf1cf0d36826e51bd4bb01fb23