150292219: Deduplicate lint classpath entries
When lint is configured with checkDependencies true,
we can end up with a large number of duplicate paths
in the classpath.

Removing the duplicate entries results in a 30% speedup
of the mean total lint time in the test project from
issue 148172219.

Combined with the patch to cache null values in UAST
that yields an additional 2%.

This CL also adds a workaround for KT-37200/150530989.

Test: Covered by existing tests
Bug: 150292219
Fixes: 150530989
Change-Id: I63c8a353ba7cf858906af640cbe53ea4b7cc2eca