Add user managed ADB server mode to ddmlib.
Ddmlib contains ADB server management logic that will kill/restart the
ADB server when connection failures are detected.  This automatic ADB
server management will also kick in when a flakey remote ADB connection
temporarily disconnects. When this happens, a local ADB server will be
started and the user won't be able to re-establish the remote ADB
connection.

This CL adds a "user managed ADB server mode" that disables ddmlib's
automatic ADB server lifecycle management.  In this mode the user will
be in full control of the ADB server.  If the server connection drops
ddmlib will simply keep trying to reconnect.  In the case of flakey
remote ADB connections, the connection will be re-established when the
server comes back online.

This new mode can be enabled by setting the environment variable:

  ANDROID_ADB_USER_MANAGED_MODE=true

Bug: 152331345
Test: Started studio in manually managed ADB server mode and setup
a remote ADB server scenario using SSH port forwarding.  Killing the SSH
tunnel no longer kills and restarts the local ADB server.  When the SSH
tunnel is re-established, studio reconnected and was able to deploy
again.

Change-Id: If6bd0d50fddf12f9dc12b0a95574ab755622129f