IRR: Continuous patch size monitoring
Problem: Some developers load large (multi-GiB) APKs into Android
Studio. This result in very large Patch being built in memory which
size is only checked at the end of the patch generation process.
Checking for size is too late to avoid an OutOfMemory Exception from
being thrown.

Solution: Monitor size of a patch as it is being constructed and
fail early if the threshold is crossed.

Fixes: 152399587
Test: Unit Tests
Change-Id: I380633c2562cdfcb2b3b99524837236bcb003a34