Switch to send absolute locations from agent to Studio.
We used to send the position relative to the parent.
In Compose all coordinates are absolute which then can be sent
without converting them to relative coordinates.

The Studio client need absolute coordinates as well.

Remove scrollX and scrollY from the protobuf since they are no
longer nedded by the Studio client.

Bug: 149999705
Test: Existing tests should pass
Change-Id: Ieb524e2168cf8f980318fc680f587bf3555268b8