Clean up DX interned tables with a build service
Stop using buildFinished callback to clean DX static state.
Instead, a build service is used which will be
closed as soon as dexing and dex merging are done
(the only two tasks using DX), and during closing
we'll clear interned data. This is used only if
DX is used so there is no overhead.

Bug: 154742527
Test: existing dx tests
Change-Id: I95cb35c2746df949963d0f4fc5b9bd5453fb4351