Capture ANDROID_AAPT_IGNORE env at configuration phase
This CL captures ANDROID_AAPT_IGNORE environment variable at task
configuration time using Gradle providerFactory.environmentVariable
api, and pass the actual value of env all the way down to the class
which use it. This means that ANDROID_AAPT_IGNORE env is considered
as an input which affects task configuration.

Bug: 148512507
Test: existing
Change-Id: Iae84c7d844e9dc85b69c45670a0054d4f78e598d