[Lint,AUA] special-case detector kind for AGP dependency
The lint check for dependencies will under the right circumstances
highlight the classpath dependency on AGP itself, and will offer to
perform a text substitution to update the AGP dependency to the latest
stable version.  Running on the command-line, this is perhaps
acceptable, but because an update to AGP might imply or even require a
number of other changes to the build, it's unhelpful to allow this
bare quickfix when running in Studio, because users end up using that
rather than Studio's support for a combined AGP version upgrade and
build file refactorings.

This change introduces a new detector kind for the AGP dependency
check, so that the Studio LintIde support can override the quick fixes
provided.

Bug: 157287524
Test: existing tests updated
Change-Id: I30fa622c36d7961078470691beed715468389c97