Zipflinger: Refactor ZipSource.select
Problem: ZipSource method select returns a Source for the sole
purpose of allowing alignment selection. While it is consistent
with the rest of the API, it is also confusing to users who may
be tempted to add the Source to a ZipArchive.

Solution: Make select method return void. Alignment is specified
otherwise, internal objects are never exposed, and adding to a
ZipArchive can only be done via add(ZipSource).

Test: NA (API update)
Bug: NA

Change-Id: I51aab3be2a6f7c0df3fe65fc55aa58f4e85b4520