[AUA] a coverage report for the upgrade assistant
A planned feature for the next release is the "AGP Upgrade Assistant",
a more comprehensive tool for users to be able to upgrade their
Android Gradle Plugin and anything else in the build of their project
that needs to change in lockstep.  A coverage report will be useful
when discussing the readiness of the feature.

Bug: 149565531
Test: none
Change-Id: I452c6dce52f428a9613ca5c4962c9999ae19053d