Add a filter to filter out potential META-INF/MANIFEST.MF in PerModuleBundleTask
In AGP 4.0 if advanced profiling in enabled, system will throw an exception in PerModuleBundleTask because
mutiple files named META-INF/MANIFEST.MF are zipped in one file. This was because the abiFilter could potentially be null
as the supportedAbis in variantDslInfo could return a null but fixed in 4.2 that it will return an empty set.
This CL is to eliminate other potential occurance of the same bug.

Bug: 140219742
Test: PerModuleBundleTaskTest
Change-Id: Ia04ecadf94de7b7d6bcf4dc4306155ae110f0cd7