New fixtures to create test projects.
Builder based mechanism to create test projects.
Starting with the HelloWorld one and converting a
user of the old one to validate things are working.

Also starting to add more coverage to the model builder
test using the new builde, including one using library
module which detected some issue on the ModelBuilder side.

Bug: N/A
Test: Updated and created new tests
Change-Id: Ia320ace3703ebb97392e92df5c7ea4f8ff61e87b