Lint: avoid redundant Kotlin compiler analyses
This improves end-to-end Lint performance on the Tachiyomi project
by ~30% in a warm Gradle daemon: Lint runs in 40s rather than 60s.

The main idea is to avoid running the Kotlin compiler in
LintDriver.runExtraPhases() if the BindingTrace from the previous
phase is still valid and applicable.

Note, this optimization only applies when checkDependencies=false,
where the set of files to analyze stays the same across phases.

To make this change possible, some things were refactored:

* The `checks` list is moved out of UastSourceList, because the source
  list is cached but the list of checks using it may change over time.

* .kts files are prepared ahead-of-time with the other UAST sources,
  in order to reduce the number of BindingTraces created/invalidated.

* In order to simplify things, UAST sources are collected
  unconditionally (regardless of the value of `scopeDetectors`).

Bug: 153758215
Test: //tools/base/lint/... and //tools/adt/idea/lint/...
Test: benchmark: //prebuilts/studio/buildbenchmarks:Tachiyomi_lint
Test: manual: sanity-checked Lint issues found in Tachiyomi project
Change-Id: I2fa3f50e294bd4b04bbb9b107d1d123731aa8ab4