Store separate build components in json files
Currently android_gradle_build.json and android_gradle_build_mini.json
stores build commands and clean commands as strings. This is
inconvenient since it requires escaping spaces when contructing it from
arguments. This change instead stores the commands as array of strings.

This change also introduces a version of intermediate files under .cxx
so that these files are regenerated when user upgrades the AGP. Also,
since the version check was not present in previous AGPs, the
build_command.txt is changed slightly (removing the unnecessary prefix
of 16 spaces) so that downgrading AGP would also trigger regeneration of
these intermediate files.

Bug: 163046035
Test: existing
Change-Id: I7736547e4e65f49131e2c5c925712508a0b6db27