[lint] checkDslPropertyAssignment cookies
checkDslPropertyAssignment takes three cookie arguments: one
corresponding to the property name; one to the property value; and one
to the whole assignment statement (which might be a method call
statement with the effect of a property assignment).

The Uast visitor correctly distinguished between all three of those,
but the Groovy visitor did not, often repeating the whole statement
for the value.  This led to inconsistencies between the reporting of
an issue in a Kotlinscript build file compared with the Groovy
equivalent.

Adjust all callers to pass three distinct cookies.  Likewise, adjust
GradleDetector to make sensible decisions about whether to highlight
the value, property, or whole statement.  Generally, in these changes,
the value is preferred except where the treatment of the value is
specific to a particular property (or dependency configuration).
Thus for example:

- generic artifact dependency suggestions highlight the dependency
  specification only;
- AGP (and other plugin) dependencies highlight the whole statement;
- accidental octal highlights only the number;
- sdkVersion properties highlight the whole statement;
- deprecated or obsolete property assignments highlight the property
  name only.

Bug: 156217625
Test: existing tests updated
Change-Id: I0aecc2677d8e0cfc3118d9ffdd1e87314cb2fc1b