[lint] highlighting for AGP dependency check
Arguably the AGP dependency should be somewhat special for the Android
linter: changing the dependency has the potential to break some IDE
services if it's not done properly, and there are checks to make sure
to offer compatible upgrades where reasonably possible (and IDE
support to handle the upgrade as a refactoring if not).

Given that, it's reasonable to highlight it slightly differently from
other dependency warnings; visually, highlighting the whole
statement (including the `classpath`) rather than just the dependency
specifier itself should give the user a clue that this is slightly
different from a generic new version suggestion.

Bug: n/a
Test: existing tests adjusted
Change-Id: If104ce3154f9a680f8bb9f54c05c2f1500201acb