Validate ANDROID_SDK_HOME & ANDROID_PREFS_ROOT together
The previous code checked one after the other, but because
one is deprecated it is better to check whether both are
set, and if that's the case, that they both have the same
value.

This will reduce confusing errors where users set one
value and see a different behavior because they don't
know the other var is also set.

Test: added new tests to AndroidLocationTest
Fixes: 165833833

Change-Id: Ifde06eff2a508d6f864a2410c20435afdad41303