un-nest VariantProperties from Variant.
it was decided that nesting onVariantProperties within the onVariants
block is very confusing to users since the two blocks are not
executed at the same time.

So instead of doing :
android {
   onVariants {
      onVariantProperties { ... }
   }
}

users will have to use the already existing other form :

android {
  onVariants { ... }
  onVariantProperties {...}
}

these APIs will continue to evolve anyhow as it is moving to the
new AndroidComponents extension.

Removed the connection from Variant to VariantProperties which
simplifies the code a lot.

Test: Existing.
Bug: N/A
Change-Id: Ie5d675515888685b42e730825e0469a59e74aab0