Initialize AnalyticsSettings before computing anonymizedProjectId
Currently, we are not getting "salt" for the anonymizedProjectId
properly because AnalyticsSettingsData instance is not initialized when
we compute the anonymizedProjectId at configuration time.

In order to fix this, this CL adds rootProjectPath as a parameter of
analytics build service, which is used to calculate project id after
AnalyticsSettings is initialized.

This issue is uncovered by b/166271208 which can only reproduced when
running multple subtests at the same time from bazel.  I suspect there
is some interaction between tests which somehow makes
AnalyticsSettingsData not null and then the system property gets
accessed without using gradle environment.

Bug: 166271208
Test: ProjectIdAnonymizationTest
Change-Id: Ie33839cf4eb6494a3144e5136910ca1928a0490b