167593732: Add SARIF as supported output format for Lint
This CL adds SARIF file as a supported output format.
It focuses on the subset of SARIF that is supported by
GitHub, plus it also adds the quickfix descriptors which
I hope will be supported in the future.

Note: Because the lint-model integration in AGP is actively
being worked on by the AGP team, this CL does not hook up
a new DSL lintOptions flag to explicitly generate SARIF
files from Gradle. Instead, the existing report methods will
emit SARIF if the file extensin is .sarif or .sarif.json,
the two file extensions specified by the SARIF spec, e.g.

  android {
    lintOptions {
      textOutput file("lint-results.sarif")
    }
  }

Fixes: 167593732
Test: Unit test included
Change-Id: I591c846ab2732789011c2103ad487f1fbc41bb7d