Specify JDK APIs when using kotlinc and javac with bazel
If JDK bootclasspath is not specified with kotlinc,
default to using JDK 8 bootclasspath and do not use
bootclasspath that kotlinc is running on. This avoids
issue of having JDK9+ APIs used in Kotlin sources, as
we still need to be able to run on JDK8 (in AGP case).

Also, add "--release 8" argument when invoking javac in
order to use OpenJDK 8 bootclasspath.

Bug: 166472930
Test: BuildActionInvokerJdk8Test
Change-Id: I2f8b3239db4b7b02e0391f17a206e23d14376b7d