Add AGP DSL for configuring Sarif
This CL updates the android.lintOptions block to
let you configure SARIF reports. This has already
been possible via a hack (use a .sarif file extension
for one of the other file formats), but this is now
deprecated (and emits a warning message), and instead
you can configure it like this:

android {
    lintOptions {
        sarifOutput file("lint-report.html")
    }
}

Or if you're okay with the default name (which also
will include the variant name), just use

android {
    lintOptions {
        sarifReport true
    }
}

This CL also makes one other change to the lint
options DSL: if you specify a file for a particular
type of report, that report will also be turned on
(currently you'd need to specify for example both
textReport true and textOutput file("my-errors.txt").

Bug: 167593732
Test: Integration tests included
Change-Id: Iefb620f3b87e4024f87ca4b2ca547ef844f151c8