Run lint in an process-isolated Gradle worker
In order to re-use between runs for performance reasons.

As Gradle has the kotlin standard library present on the classpath even
in the process-isolated worker case, create a separate classloader
within the isolated process that is kept in a static field.

Disposes of the application environment after each lint invocation
for the moment to avoid locking files.

Also adds flag to run lint in the Android Gradle Plugin
process for debugging purposes.

Bug: 160392650
Test: Existing (covers the process the
      in-process path is only test manually but is not intended
      for external consumption, it's primarily for debugging.)
Change-Id: If9de736d454588c6a13ed03869c208e8543ed238