Add some missing functionality to DelegatingPath
Kotlin only delegates methods defined or overridden in the
interface being delegated, so toString, equals, and hashCode
need to be handled explicitly.
endsWith, startsWith, and compareTo need to compare the delegates of the
other path.
We need to declare @Throws(IOException) if subclasses are going to be
able to throw.

Test: this is a test change
Bug: none
Change-Id: I60b83fa4504bdd17b9f23f1c21e27d9133269c4a