IRR: Externalize Sites lib
Problem: Java Sites must be used both on host and device.
Solution: Move Sites to its own lib

Rational: This solution (#1) has the inconvenience of adding a jar
to the build process and released package. There would be a different
approach (#2) in using a static Sites.java as source of truth which
generates a .h/.c. I find #1 > #2 because:

- #1 relies on '+' concatenator common to C++/Java instead of
  #2's %s. This result in less complicated code to maintain (%s
  would involve sprintf, buffer malloc/free, and overflow avoidance.
- #2 "hides" dependencies of the .h/.cc on Sites.java in the build
  system. A dev modifying Sites.java could break the .cc (by changing
  the %s pattern and not even know it). #1 makes is clear that both
  Java and C++ code is generated.

Bug: NA
Test: NA
Change-Id: Id10767034908ec026d2f7a9e0eb2a5f1e3fe57d1