Fix prefab publishing for hyphenated modules.
Cases where the there were multiple modules that shared a prefix up to
the first hyphen were being wrongly matched by this bad assumption. We
do have the proper module name to match instead, so use that instead
of this strange matching behavior.

The case that was reported was a project with two modules: `settings`
and `settings-native-library`. `settings` was supposed to be published
in the AAR, but because of the hyphen split both modules were being
identified as `settings`.

Test: added regression test
Bug: https://issuetracker.google.com/173348031
Change-Id: I8412fd44c4636445c52f3985f1a42d484b119359