Get rid of LintDetectorTest.TestFile
Lint had two separate "TestFile" classes, and for backwards
compatibility back in the day I made one extend the other.  However, I
never went back and cleaned this up, and as a result all the APIs
require use of the older one (which extends the newer one).

This CL removes the nested TestFile. This is mostly compatible for unit
tests that simply use the lint() DSL, but for any test which explicitly
declares test files (for example to reuse them) this will require an
additional import (which the IDE can now unambigiously suggest).

Test: Existing
Bug: None
Change-Id: Iab87621e8973bd2a8cec356b5180531666607722