removed onUnitTests and onAndroidTests methods.
Instead of providing explicit methods to access the unit test
and android test variants, users will now be able to use the
single onVariants API and access the test variants from the passed
variant interface.

The unitTest and androidTest methods on Variant allow to access
these test variants. The methods return nullable types in case
the test variant is disabled (in the VariantBuilder for instance).

Kept the old onUnitTests and onAndroidTests methods until beta6
so androidx and other users have time to switch to the new
API.

todo: new APIs/interfaces to be reviewed in API Council

Test: Added unit and API integration tests
Bug: 177504396
Change-Id: I3136bad49c1e0625c60a98c527c4f0dad892fce9