Do not overwrite additionalOutputDir if it is already set
Previously AGP would overwrite additionalOutputDir even if it is set to
guarantee that the provided folder is readable and writable to by
instrumented test code. However, this use-case does not support
Macrobenchmark and other test infra that would like to customize the
folder to one with more relaxed access restrictions.

This also provides a way to have a unified code path which handles copying
test output into a single well known dir after all tests have completed,
but before test apk is uninstalled.

Fixes: 181818399
Test: //tools/base/build-system/builder:tests.test
Change-Id: I17d48329120a19c1444f122474dbe13c94b01a60