Fix CMake locator perf issue
This is something I found while working on the timer class. CMake locator is supposed to have a special case for fork CMake versions (build.gradle = 3.6.0, SDK = 3.6.4111459). The locator was not properly comparing version numbers in the path that physically looks at the SDK folder for matching versions. It's technically a bug not just a perf issue but it happened to be covered by downstream (but slower) checks.

Bug: NA
Test: Unittested
Change-Id: I6a8444160277ddb5f109fdcc2da7724c59fb3abb