Temp workaround on Crostini to kill accessibility daemons
There is supposed to be only one instance of these daemons running. When
the second one spawns, the OS should have killed it off immediately. But
the Crostini does not handle this properly. Each test execution triggers
new set of processes, and ultimately causes OOM failures after ~2 weeks
of running builds and tests. This is a temp workaround to kill these off
manually.
Bug: 159371003, 181588403
Test: tested locally

Change-Id: Idbe086c5b28ee04e899bf0965d8d2f31ee1ad691