Modify maven_repo rule for RepoLinker library
This change adds an alternate rule to _maven_repo, _maven_repo_list.
Instead of bundling artifacts in a zip, it generates a CSV file that
contains paths to artifacts that should be part of the repository.

The separator used to delimit the path and classifier is changed to a
comma for better compatibility with Windows paths.

Bug: 148081564
Test: Existing tests pass
Change-Id: If4406fac744e0eea6181416b67d49a3e3c59e68f