Work around for pom packaging artifact groovy-all
This blocks upgrading the groovy-all artifact,
which has moved from being a jar with all the dependencies
compiled in to it to a pom packaging artifact that declares
its dependencies.

Gradle will automatically just read the pom (and not the jar) and
include the transitive dependencies. For now just special case the
groovy-all artifact to request the pom.

Bug: 181947044
Bug: 183921630
Test: Manual for now, exercised with the groovy upgrade (next change)
Change-Id: I1633f40c61a7960d39a1265777f42400265d876f