[AGP Tooling Model] Fix handling of API jars
When an api.jar is present in the AAR, that is considered the entire API
of the AAR by the build system, replacing classes.jar and any jars in
the libs/ directory.

Before this change, this was not reflected correctly in the v2 model the
libs/ jars were included unconditionally.

This change makes the new model v2 reflect that, and fixes a
transcription error in the location of that jar, as well as sorting the
jars by name to make the model deterministic.

A separate change will be needed to make Studio's interpretation of the
v1 model correct.

Bug: 185363193
Test: Added new AarApiJarModelTest and infrastructure for it.
Change-Id: I2bc1b7b5b52cb8300b0e5700f4180895fefb59f7