Replace symlinks rather than writing when unzipping
When run locally, unzipping Maven repos that contain artifacts that have been linked with RepoLinker causes InstallerUtil.unzip to try to write into the original file. Because the original file is in the Bazel sandbox, it causes an AccessDeniedException to be thrown.

This change changes InstallerUtil.unzip behavior to replace the symlink instead when it encounters one.

Bug: 185775835
Test: Manual test
Change-Id: Iac350fc99b4acf23b058431a05b37a8440bf215a