Fix decorator when gradle-api is used in buildSrc
The DSL decorator defines subclasses of DSL types with implementations
of abstract methods to make the DSL more uniform and implement locking.

The decorator defines classes at runtime using
MethodHandles.Lookup.defineClass, defining the implementation class in
the same classloader as the DSL class.

In some cases, no custom implementation is needed, and initially this
meant that those implementations would be defined in the same
classloader as the gradle-api classes, and those implementations refer
to types defined in gradle-core. This is ok in the general case, as
normally gradle-core and gradle-api are in the same classloader.

However, in cases with buildSrc this can mean that gradle-api classes
are in a parent classloader from gradle-core classes. Then the generated
classes refer to types in gradle-core, which are not visible from their
classloader. The defineClass API needs a related class, so even for
types with no custom implementation, define an 'Impl' interface in
gradle-core and decorate that instead.

Updated SupportedPropertyTypeTest to assert that the implementation
types are in the gradle-core package.

Bug: N/A
Test: Expanded SupportedPropertyTypeTest
Test: gradle-core apitTest with 8cec152f25de8156b34edc393a10422cc438a366 reverted
Change-Id: I41986e2d2a9cf747dfff0ca4426ae6f99ab50a95