Add AndroidLintTextOutputTask
Previously, if the AndroidLintTask was UP-TO-DATE, AGP would not print
any lint issues when the user ran `./gradlew lint` (which wrongly
suggests that there are no lint issues).  With this change, AGP prints
the appropriate text output even if the AndroidLintTask is UP-TO-DATE.
To achieve this, this change adds a new AndroidLintTextOutputTask task,
which is never UP-TO-DATE. This task (1) prints any necessary text
output, and (2) fails the build if there are lint errors.

This change also prevents AGP from printing lint output if it starts
with "No issues found" or if it contains "0 errors, 0 warnings".

Fixes: 191897708
Fixes: 158259845
Test: LintTextOutputTest, AndroidLintTextOutputTaskTest
Change-Id: I002903f8539345f832518f422fe7072de7de54b4