Handle module dependencies with no precompiled resource deps
When running process<variant>AndroidTestResources from a module with a
dependency on a module with no precompiled dependencies, linking will
fail when the library's intermediates/compiled_local_resources directory
is attemped to be added as a resource directory; because it doesn't
exist. This change will not add the precompiled resource directory
if it does not exist i.e the AarResourcesCompilerTransform has not been
executed.

Bug: 192441391
Test: works locally, PrecompileRemoteResourcesTest
Change-Id: Ied2bdf71919307ea92b92ca420f745e79611b6db