Provide bootClasspath through the Variant API.
Adding bootClasspath to the Variant API and return a
Provider<List<RegularFile>>.

The Provider is protected from reading during configuration time
and should only be used at execution phase (as an Input most likely).

Required all internal calls to the Variant API version (through
the usual CreationConfig interfaces). Removed the API on the
VariantScope, kept the one on the GlobalScope but it should be
moved to a dedicated class handling classpaths calculation in a
subsequent CL.

Test: Existing.
Bug: 191149720
Change-Id: Icf34f6933bd1cc521315f23fe0228f87ee9ee775