Fix transitive deps support
In previous CL, we used java_exports to add transitive dependencies.
This, however, creates a new, empty jar which depends on the
original jar. Bazel consumers can depend on the empty jar, and
access the transitive dependency. But, the produced jars are incorrect,
and cannot be published. Only users that depend on _bzl targets can
get the correct jar.

If we can guarantee that only _bzl targets will be consumed (e.g.,
by hybrid_gmaven), then we don't need this CL.

Bug: 190268212
Test: check contents of generated jars for old rules
Change-Id: Ic8bd6401da7dd17082e33e358a989610fcfb0e1b