Fixes issues with the generator
Fixes a bug where non-imported rules where taking the file
type into consideration, but it does not matter.

Fixes problem where if the generator runs, and triggers
aether "downloading" maven-metadata.xml from the file://remote
to the local repo under the name maven-metadata-prebuilts.xml.
If you do a sync at this moment or a new MAVEN_FETCH and a
maven-metadata.xml changes, then aether will used the cached
version of the metadata for a while.

Test: N/A
Bug: N/A
Change-Id: I37b1687b31bd28578a03ca28ed9e6cfdf5e5571f