Normalizes link destination in RepoLinker
The new external repo paths look like:

...runfiles/__main__/../maven_tests/path/to/pom

which on Windows seem to have a peculiar behaviour.
While I cannot find documentation for this, experiments
show that when the path contains a .. value, it is then
limited to the 260 max path length, causing issues.

This change normalizes the link target path, fixing the
issue.

Test: existing tests don't work without this and maven_tests
Bug: 190268212
Change-Id: I3174e4fa71e8d261c9106044afe94226c9e9c1a0