113122598: visitAnnotationUsage does not visit fields or static fields
First, this CL updates the annotation mechanism to invoke the annotation
callback on field references where a containing class or package is
annotated with a relevant annotation. These are assigned different
annotation usage types, so detectors can decide whether they apply to
these as well.

It also updates the annotation handler to support annotated outer
classes (which wasn't supported for methods either).

It also updates the annotation handler to support class literals and
object literals.

Finally, it fixes the resolve symbol checker to also work for fields
and default methods.

Fixes: 113122598
Fixes: 185115302
Test: New unit tests included
Change-Id: I6e270b04ca6538a9d1de8c99754aeb0bc25c171d