Skip binary files in test failure output
If the source-transforming test mode comparison fails, lint
will emit not just the diff of the error messages, but it
includes all the modified test files as well. This had two
problems: first, the path did not include the actual file
name, only the containing directory, and it included binary
files. This CL fixes both.

Test: Manual
Bug: None
Change-Id: Ic0c168416ae6cbf43b0260aaa04ea5b4a7817df2