197146610: Filter issues by platform during reporting, not analysis
When we analyze a module, we don't know the platform type of the
reporting module, and it can vary. For example, we may have a utility
Java module which is consumed both from an Android app module as well as
a server module.

Because of this, when we analyze a module, we have to assume all
platforms are eligible, and include everything in the partial results.
Then, when generating a report (and we know the intended platform), we
can filter out incidents that are not applicable.

Fixes: 197146610
Test: Integration test included
Change-Id: Ia6f46f1c423adc55b4923ef11674feb44f08909a