Remove lint's usage of gradle.user.home when invoked from AGP
Previously, lint looked in the gradle.user.home directory to check for
cached versions of dependencies, but this behavior isn't compatible with
making the lint analysis task cacheable (or able to be UP-TO-DATE) in
AGP.

An alternative approach would be to make the contents of the
<gradle.user.home>/caches/modules-2/files-2.1/ directory an input to the
lint analysis task, but that would drastically reduce the number of
cache hits.

This change also removes javaHome from EnvironmentVariableInputs; the
javaHome input was always unnecessary because AGP passes this
information to lint via the --jdk-home flag.

Bug: 183998131
Test: modified existing
Change-Id: Ia0fd3e7b7e24d1a903a3b4da5eab836f9f6d76b1