[Hybrid build] Lint is compiled with bazel
For now lint-gradle is still compiled with gradle,
as it depends on builder, which contains TypeDefRemover

The current plan is to move TypeDefRemover in to its own lint libs artifact,
built with bazel, and cut the dependency of lint-gradle on AGP.

Then lint-gradle can be built in bazel along with the rest of lint.

Then the AGP gradle build will only depend on lint-model and lint-typedefremover,
allowing parallel compilation of both in the hybrid build.

Bug: 143864616
Test: Existing tests including GmavenZipTest
Change-Id: I56060ae79a8af70a420ec0c93938b5b645be122c