Revert "Use testFixtures variant name for testFixtures resolved artifacts"
This reverts commit 5ef7a343355003835714e7938a7ebd412e4527ff.

Reason for revert: variantAttr is no longer used in v2 model, however
in v1, we still need to select the correct variant in case of
testFixtures module dependency. And since testFixtures is not
supported in v1 models, it's better to set the variantAttr to the
main variant instead of the testFixtures variant so the IDE could
pick up the main variant directly.

Change-Id: I0e2613b76ec2f6076201c8256d3c109923476322