Revert "Add a sourceset field to LintModelModuleLibrary"
This reverts commit ca7b80b6e569bb6d4cbda3b9b2b5c14e3a8a105b.

Reason for revert: Moving from using a sourceset field to rely on the
artifact address to distinguish between different sourceset artifacts.
There is no need currently for lint to know whether a library or a
module is a testFixtures component of a library or a module. By
moving to using the artifact address lint will deal with them as
separate artifacts.

Change-Id: Id2b5cdee0d13c36d15d53b85b9674329c6c6403b