Fix asset pack matching to manifest file
To match asset pack manifests to their names, the task was doing a
simple path.contains(name), which would match anything in the path,
including the names of folders that have nothing to do with the project
and partial names, like assets1 and assets10. This change makes the
matching happen between the actual folder where the asset pack build
folder is located and the path to the manifest file.

Fixes: 201263841
Test: AssetPackTest
Change-Id: I55cead98c00dbc5f24b718faa52fa2e18ef5e028