Refactor file-based properties protos
The main goal of this refactor is to help make the code more clear when
importing the variant messages.

* Remove references to "sync", since we're now targeting more
information than only what we used to produce during sync like the list
of all application ids of a module. Since this file-based properties
logic will ideally be well isolated from sync (while both contributes to
the Android/ProjectSystem models), I believe this will also help future
discussions around this feature.

* Rename some messages to allow more clear code, like for example the
top level message Variant to VariantProperties, which make it more
obvious when reading code that reads and have to process that message in
the IDE side.

* Change java_package to com.android.ide.common.build.filebasedproperties
prefix and put each proto on its own subpackage. This will make imports
longer but more organized and clear what exactly we're dealing with.

Bug: 204978506
Test: existing
Change-Id: I28f31ed1b636a260232d074c3f821e816323af38