[SDK Index] Add GooglePlaySdkIndex
... so it can be used in GradleDetector. It is responsible of loading
the information from the proto snapshot, downloading new versions and
make the information available to GradleDetector and other components.
For users that work offline, an initial snapshot of the proto index is
stored as a resource; GooglePlaySdkIndex will try to load the index in
this order:

1.- From network if cached snapshot is too old and network is available.
2.- Cached napshot.
3.- Snapshot from resource.

This change also removes DeprecatedSdkRegistry since it is no longer
needed and changes a couple of lint issues to match severities discussed
during UX research.

The included snapshot contains the information from the index as of
March 14, 2022.

Bug: 219998659
Test: GradleDetectorTest, GooglePlaySdkIndexTest
Change-Id: I685433cb6f0efe30899a912b9eef4b3f8830de82