Require Kotlin plugin 1.5.20
In earlier versions, kapt would resolve all its inputs during
configuration, requiring all inputs to the annotation processor to be
merged in a task. (e.g. https://issuetracker.google.com/183423660)

Databinding relies on many extra inputs to kapt, and this regressed
again with the introduction of the 'namespace' DSL, meaning that
databinding is unusable with kotlin versions below 1.5.20

There are now three such tasks, and this will allow them to be removed,
and the build simplified

Bug: 221250431
Test: Clean up to follow
Change-Id: Iad812c89b97645beb90c7ddea9018e8a85593873