Update LiveEdit proxies to handle suspend lambdas
When working with coroutines/suspend functions, Kotlin generates lambda
classes that extend from SuspendLambda/RestrictedSuspendLambda. These
classes implement much of the boilerplate logic to make
coroutines/suspend function properly.

In order for LiveEdit to support coroutines, the proxy objects that
LiveEdit generates to represent lambda objects must also implement this
logic.

To handle this, generated proxy objects that extend Lambda,
SuspendLambda, or RestrictedSuspendLambda create an instance of their
superclass on creation. Access to methods or fields implemented by
the superclass are forwarded to that object.

When a new proxy class is created in LiveEdit, its bytecode is parsed by
ASM to determine its superclass (if any) and implemented interfaces.
These are used to determine the transitive closure of the supers of the
proxied class. Methods and fields belonging to any of these objects that
are not overriden in the proxied class bytecode are forwarded to the
super object. The proxied object may also be cast to any of the types in
its transitive closure, or the type it represents.

This CL also moves the LiveEdit dex from the bootstrap class loader to
the application class loader. This is required in order for LiveEdit to
access the Lambda* classes, which are present in the app class loader.
Access to these classes from the bootstrap class loader is denied by
ART.

Fixes: 219538581
Test: LambdasSuperTest.java fully exercises both a suspend coroutine and a
restricted suspend coroutine.
Change-Id: I9d073cb1b740985e06f87be4cea80897e1596728

Change-Id: Ie93a2b6a51277fe0deb39f4507d0e263c42526f2