Softly enforce `android.enableDexingArtifactTransform` feature
The feature has been enabled by default for a few years with 0.81% of
users using the flag in the last week, so we will now SOFTLY-ENFORCE it
in AGP 7.3 and ENFORCE it in AGP 8.0.

The same for `android.enableDexingArtifactTransform.desugaring` and
`android.enableDexingArtifactTransformForExternalLibs`.

NOTE: As mentioned at bug 225900051, we'll still need to watch out for a
few cases where people might be blocked if the feature is enforced.

Test: Updated BooleanOptionTest
Bug: 225900051
Change-Id: I4a188b368545f9f32682678a831a4c9a12ecfd3d