Add Variant.allComponents
Variant.allComponents is similar to the existing
Variant.nestedComponents property; the difference is that the variant
itself is also included in the Variant.allComponents list.

This change also adds the the following functions to Variant:

    fun nestedComponents(action: (Component) -> Unit)

    fun allComponents(action: (Component) -> Unit)

The main motivation behind this change is to provide a simple
alternative to Gradle's inefficient Configurations.all {}. With this
change, a user can replace Configurations.all {} with the following:

androidComponents {
    onVariants { variant ->
        variant.allComponents { component ->
            component.configurations {}
        }
    }
}

In a subsequent change, we'll revisit Component.configurations() to
offer more control over exactly which configurations are affected.

Bug: 72477422
Test: ConfigurationsTest, AnalyticsEnabledVariantTest
Change-Id: Ic93b01f82a323a53cd678300cd782a864bcee983