Add foreground_process_tracker library
This change adds a new `foreground_process_tracker` library.
This library is responsible for tracking changes of the
foreground process on the device.

This is achieved by using the dumpsys activity processes
command. From the output of the command foreground process
PID and package name are extracted.

Once a change in foreground process is detected, a
LAYOUT_INSPECTOR_FOREGROUND_PROCESS event is sent to Studio.

The library starts checking for changes only when the
START_TRACKING_FOREGROUND_PROCESS command is received from
Studio.

The library stops checking for changes when the
STOP_TRACKING_FOREGROUND_PROCESS command is received from
Studio. Or when the adb shell process is terminated.

The tests use fakeandroid. They send commands to the library
to start and stop tracking the foreground process.

A mock version of the transport library is created.
The mock library can take mock dependencies,
like a mock implementation of the foreground_process_tracker,
which uses a mock BashCommandRunner to avoid calling dumpsys,
which is not available in fakeandroid.

Test: dynamic-layout-inspector/foreground-process-detection/test/ForegroundProcessTrackerTest.java
Bug: 232067042
Change-Id: I8e9aded96646012a05deaa2018102d7c1bfcb04d