Fix incorrect test reporting for split test targets
With the previous usage of --test_filter, when targets hit the bazel
timeout they would report all tests as having failed, not just the
filtered ones. Instead of relying on --test_filter, the
JarTestSuiteRunner performs the filtering based on system properties.

The incorrect reporting of which tests failed is a problem when tests
are reported using go/antswatcher. Since a test fails in one attempt
(the timeout) but is never run in the successful attempt, this is
considered a test error in AnTS. To have this register as a flaky test,
we need the tests reported to be accurate during timeout failures.

Bug: 173826673
Test: manually verified tests in split targets
Change-Id: Ic558447e34bff2a8aeac838fd30eec9a54c623ec