adblib: Fix flaky test
Using a `synchronizeList` for the list of events in
`TestDeviceClientManagerListener`, and exposing that list
as a property introduced a flaky test behavior: if the list of events
is modified while a consumer iterates the list, tests would sometime
result in throwing `java.util.ConcurrentModificationException`.

The fix is to use custom synchronization, and make sure to
either modify/filter the list in a synchronized block or make
a copy of the whole list to make it easier to write tests.

The performance impact is minimal because this is only
for test code, and there are very few events in current
tests.

Bug: 188067646
Test: This is it
Change-Id: I8e7254b03708dc367aef0513bdc8191cece28dfb