Switch back to Disconnected state on connection timeout
When connection or disconnection times out, we need to cancel the job
trying to connect or disconnect and return to the original state.

This is accomplished with a utility method that handles the atomic state
updates required to do this without overwriting concurrent updates to
the state.

This approach eliminates the need for extra state in the Activating /
Deactivating state objects.

Bug: 241544067
Test: added
Change-Id: I52e0a594b4a52693c2ade8555c7b9e623cdb0c64