Allow @Suppress to apply to getters and setters too
If you have
  @Suppress("id") var foo: Foo
this will only suppress issues flagged on the private
backing field (UField), not the generated setters and
getters for this property. This CL makes it work for
those as well. This isn't how annotations work in Kotlin,
but we already take some liberties with @Suppress to
make it work in a more intuitive and flexible way.

Bug: Not filed
Test: Unit test included
Change-Id: Ibc34ca366e0c1994c806d064ea736e81f42814b9