Generate correct databinding package import
All viewbinding classes are generated under the root project package with `databinding` appended. But when an activity (or other recipe) is created outside of the project's root package, the import statement was incorrectly pointing towards the sub-package rather than the root package.

This change updates all the calls to this import statement to use the
root application package instead.

The "interesting" part of the change is in viewBindingUtils.kt. The
rest of the change is effectively just passing various parameters
around.

Test: TemplateTest.kt (in separate linked cl)
Fixes: 200894865
Change-Id: Ie11bff7c99c5c9b15fbfa581f76e48891f179776