Separate the logic that does the capturing from ViewLayoutInspector
This change makes the capturing logic more explicit.

Currently the capturing code is hidden behind a generic Executor,
which runs only when the screenshot is captured.
Currently is very hard to understand that capturing a bitmap screenshot
has the side effect of making everything work (send events to studio,
capture views information etc).

Using a specific Exceutor, is a step towards decoupling screenshot
capturing and the rest of Layout Inspector functionalities.

Test: existing
Bug: NA
Change-Id: I77826b59f93068fd9c09390f38229618abf85167