Rename ProfilingApp to CaptureInfo
To help unify the trace-manager for both
the memory and cpu profilers, the ProfilingApp
struct has been renamed to reflect a more generalized
structure for capture information.

The StartProfiling and StopProfiling methods of
the TraceManager have also been renamed to
StartCapture and StopCapture respectively to
avoid the ambiguity of the term "Profiling".

Bug: None
Test: existing
Change-Id: Ida4b9d69f3471bb408392674681879254321667f