Fix comparison of Kotlin PSI variables in AnnotationHandler
Now correctly compares only the javaPsi of Kotlin UVariables when searching for names which resolve to the same annotated variable, instead of the sourcePsi. This fixes an issue where visitAnnotationUsage may not be called back in some cases, e.g. b/247025215. However, NO other existing detectors are affected, as this situation is fairly niche: this issue manifests if and only if a detector wants to visit usages of an annotated local variable other than method calls, array accesses, or unary operators. These cases are handled separately. In the case of b/247025215, we wanted to visit usages at binary comparison operators.

Bug: 247025215
Test: testComparingResourceTypesKotlin
Change-Id: Ie1c06521d233975ea90b59200aa59da202e872a5