Add utility for converting Java TimeUnit to DurationUnit
This method exists in the Kotlin stdlib but is currently experimental.
This is a drop-in replacement we can remove when the API becomes
stable.

Note that Duration and DurationUnit are no longer experimental as of
Kotlin 1.6 but for some reason we still need the opt-in annotation
for IJ to be happy.

Bug: None
Test: DurationUtilTest.kt
Change-Id: Ib25d38f2ada0a1a5efc50fc761f0486d393cee75