Add Artifact API to include external debug symbols
This change add a new API to add a new artifact to a multiple artifact type. With this API an existing directory with debug symbols can be added to a newly added artifact type. MergeNativeDebugMetadataTask consumes this artifact to get all the directories with externally sourced debug symbols. These debug symbols are merged with the debug symbols extracted from the unstripped native libraries.

Bug: 208593320
Test: ArtifactsImplTest, MergeNativeDebugMetadataTask
Change-Id: Icf30d403d1d4024bb71eabb9058a73b6447ac0bb