Adapt trace manager to unified trace config
To take advantage of the new unified TraceConfiguration
proto (and its exclsuive studio-side creation), the
StartCapture method has been adapted to read this new
unified config proto. It still allows for utilizing the
old UserOptions-based TraceConfiguration by using a
boolean parameter to choose between the new and old
config protos.

To verify the trace mananger still works with the new
unified trace config, a parameterized test suite was
created to test the trace manager with the old and new
configuration protos.

Bug: 258515754
Test: added
Change-Id: Ibc0d8d431ad75f558545efc1d948a1e914cadbe0