Don't map ListProperty hoping for task's dependency
The LayeredSourceDirectoriesImpl uses a ListProperty to book keep
all instances of DirectoryEntries. DirectoryEntries themselves can
have DirectoryEntry that have a Provider with task dependency embedded.
Mapping on the ListProperty does not carry the Task's dependency, as
Gradle probably stop at the first value object.

Instead of returning a Provider<> which is misleading since users
could expect to be able to map its value and get proper task dependency,
the API will now return the List<DirectoryEntries> directly. This
makes it very clear that you cannot map its value.

There were a fair number of the API that were doing map() and get()
so all got reworked to deal with the List<> instead.

Test: integration-test
Bug: 261329823
Change-Id: I2cbd979d8165291e190730543bf9da42e55420e2