[li-overlay] Make LayoutEvent#screenshot optional
This change is in preparation of the integration between
LayoutInspector and RunningDevices.
We won't need bitmap screenshots after the integration, but there
is going to be a transition period where both old and new
LayoutInspector will need to work. Which will require the
ability to switch bitmap collection on and off.

This change adds a `disableBitmapScreenshot` property to the
InspectorState.

When set to true, this property prevents the agent from capturing
and sending bitmap screenshot to Studio.

The change does not affect SKP images, as we will still need them
to capture snapshots.

The property can be set from Studio using
DisableBitmapScreenshotCommand.

Test: ViewLayoutInspectorTest
Bug: 265150325
Change-Id: I6c7fa6ffeca8810c0ddbe69891329dff5029ddb6