JDWP-Packets: Introduce new concepts for scache
Fundation work for "scache" library which needs to not only parse
but also write packets ahead of time.

Introduce Packetable, Cmd, and Reply to allow scache to synthetize
packets

Introduce CmdSet and Cmds which will used to replace literal
declarations in libjdwp-tracer.

Introduce the concepts of Key (to enable packet caching in scache),
and packet type (cmdset-cmd).

Test: test/*
Bug: NA
Change-Id: I50553a6518e8ec812ef7170f2a75e9dc2ba85919