Add test for desugaring nio apis
java.nio is a new set of apis that library desugaring supports. There are
three versions of desugar library dependency: basic, minimal and nio. Nio
version is the superset of basic version which includes nio apis.

This CL also deleted a disabled test in SeparateTestModuleWithMinifiedAppTest
because we don't do obfuscation for dexing in debug mode anymore.

Bug: n/a
Test: CoreLibraryDesugarTest
Change-Id: I7b066870245f72c922d39dd3e9de1d5c21c7cd66