Define a method for the inverse of parsing a VersionRange
We will want to know, in subsequent changes, whether a given
VersionRange is externalizeable as a fragment of Gradle dependency
syntax.  For that, toString() is not suitable, as that must always
return a String no matter what.  Define toIdentifier() and use it to
separate out cases where we will be able to reparse the String from
cases where we won't.

Bug: 242691473
Test: added
Change-Id: I237e0835fdbcbdb3dca71c490404e5a99a33d489