Fix handling of android:extractNativeLibs in manifest merger
Previously, if an Android library dependency had
android:extractNativeLibs in its AndroidManifest.xml, that attribute
would be merged into a downstream app's merged manifest.

With this change, the manifest merger does not merge
android:extractNativeLibs from dependencies. Instead, if a dependency
has android:extractNativeLibs set to true but the app's merged manifest
does not have it set to true, there will be a warning (which can be
suppressed by adding
android.experimental.suppressExtractNativeLibsWarnings=true to the
gradle.properties file).

Bug: 272553504
Test: ProcessApplicationManifestTest, ManifestMerger2SmallTest, etc.
Change-Id: Ie7407c3af2274c1405b28fd94da639a0fa793287