Use Different Golden Dir Per Engine Type
Prepare for the Android Engine that will need to write to a different
dir. Note that SIMPLE & JVM have exactly the same results but only one
of them will actually remain by the end of this project.

It's not yet clear if we need DEX on local host & Android or just one of
them. It's also not yet clear if they will also have the same result.

I also included a small change to the tests so that they share more
code. Note that it's probably possible to combine them all in a single
class, but I think there's actually value in having them separate.
Mainly during development to make it easy to focus on one specific type
at a time.

Bug: NA
Test: NA
Change-Id: I1bd474211f75438df429a43835ec7938efa7286a