Decrease the default amount of memory earmarked for lint tasks
The amount of memory earmarked for lint tasks was recently increased
from 512mb to 1024mb. This change allows users to set how much memory to
reserve per lint task and decreases the default back to 512mb.

The motivation for this decrease is the plan to move to per component
lint analysis.

Bug: 197322928
Test: LintParallelBuildServiceTest
Change-Id: I7cc608bc12d5490509770bfd425cf631e5d6e835