Do lint analysis per component when using standalone lint plugin
Previously, the main and test components were analyzed in the same lint
invocation.

With this change, test sources are analyzed in a separate task (1) if
android.experimental.lint.analysisPerComponent is true or (2) if it's a
KMP project.

This change also creates a separate lint cache directory per task
(whether or not doing lint analysis per component) to avoid concurrency
issues when sharing lint cache directories between tasks.

Bug: 197322928
Test: LintStandaloneTest
Change-Id: I9a49bd33f3b7daa8ea844bc9d0d6535839499d83