adblib: Address potential deadlock
Fix a potential deadlock between the "activation" and "receiver"
coroutines in `SharedJdwpSessionImpl`: If a receiver
waits (e.g. using a `CompletableDeferred`) for its
activation to complete, this may lead to a deadlock
in `SharedJdwpSessionImpl`. The root cause is that
the "activation" coroutine can block on
`SharedFlow.emit()` while sending the `StartPacket`,
so the solution is to make sure the "Receiver" coroutine
does very little work when processing the `StartPacket`,
including *not* calling the `receive` callback (user code).

Bug: 188067646
Test: Included
Change-Id: I24135eb1363d2b41ab3370c71a2e63866703605f