adblib: Delay process close notifications
Due to the asynchronous nature of process tracking,
`AppProcessTracker` and `JdwpProcessTracker` sometimes
get notified of a process termination before the
corresponding JDWP connection receives an EOF from
the device. This can lead to a cascading effect which
results in a `CancellationException` of a `JdwpReceiver`,
instead of a normal termination.

The workaround is to delay closing `AppProcess` and
`JdwpProcess` by a few milliseconds (100 by default)
to prevent the `CancellationException` from being
thrown.

Bug: 291272975
Test: Fixes flakiness of AdbLibDeviceClientManagerTest.testClientKillWorks
Change-Id: I99fd340e3caa7d51f11794e5c6c26d09acc786a1