Handle type annotation annotations in AnnotationHandler
Note that we now show *two* warnings for a line like this:

val x = AnnotatedClass()

one for the constructor call and one for the implicit type
reference assigning the type AnnotatedClass to the variable x.
Since the type reference in this case isn't present in the original
source code, the location of the warning gets expanded to the entire
line.

Fixes: 228961124
Bug: 286616170
Test: AnnotationHandlerTest.testTypeReferences
Change-Id: I7e98640b6d578b483c4c2069cf4ce0def32f55e7