adblib: add support for root/unroot
Add support for `root` and `unRoot` device services.

Also add the ability to "clone" any `DeviceSelector` into a
similar `DeviceSelector` than will capture a transport ID
when invoking an ADB service. This is useful to callers of `root`
and `unRoot` to be able to wait for a device to become "offline"
after restarting.

Bug: 188067646
Test: Included
Change-Id: If15169cb23dc7637a671397df188b65db43a3880