Move manifest file existence check to task action
Previously, we performed the file existence check lazily inside a
Provider, and we expected Gradle to resolve the Provider at task
execution time. However, Gradle actually resolves the Provider during
task graph creation, which means the file existence check will be part
of the configuration cache inputs.

To reduce the number of configuration cache inputs, we should move the
file existence check to inside the task action.

(This was also partially done by a recent commit 2374400.)

Bug: 289098893
Test: Updated ConfigurationCacheReportChecker
Change-Id: Ic1cd744facc368c59cf713095459c1dee9ccff4f