Handle recursive calls in potential version checks
The code which tries to figure out if an API call with a version
requirement is surrounded by a version check handles indirection.

However, this code could get tripped up by some types of recursive
methods which would lead to a stack overflow exception. This is highly
unlikely to be the case for a version check utility, but the code also
gets called on methods that turn out not to be version check utilities,
and we need to gracefully handle this.

Fixes: 290340814 - Unexpected failure during lint analysis of java file
Test: Unit tests included
Change-Id: I55d2dbf5db892e8639668926d187da09ca9b0c38