Do not use implicit task dependency for default proguard files.
For compatibility reasons, the `consumerProguardFiles` file collection contains
the default proguard files that can be specified in the DSL using :
proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"))
This is problematic because such default proguard files are actually extracted
from the AGP jars resources. The extraction is done by `ExtractProguardFiles` task
and stored in the InternalArtifactType.DEFAULT_PROGUARD_FILES.
Since `consumerProguardFiles` is a List<File>, we cannot represent the task
dependency.
Therefore even though the file reference is contained in `consumerProguardFiles`,
it represents an implicit binding that is flagged by Gradle as an Error. To avoid
the problem, we should remove all the default proguard files from the
`consumerProguardFiles` collection and use a normal @InputFiles annotated field
to store the DEFAULT_PROGUARD_FILES artifact.

Also added support to create integration tests with kts files rather than
groovy.

Test: Added regression tes
Bug: 295666695
Change-Id: Ica7d8b5acd19137d523f9be0e29fd2f8d0b5fc56