Don't Send Consecutive Zero Speed Events
Currently, the agent sends a speed event every 500ms. Even if there is no traffic

This results in a lot of unnecessary communication between the agent and Studio

It also used to result in a lot of unnecessary speed data stored in Studio. I fixed that on Studio side already but this is a better fix

Instead of sending an event every 500ms, the new code will only send an event if it's non trivial, as in, contains values that are not zero.
Unless of course if it's a boundary event, as in, right before or right after a non zero event

So for example, a series of `00035700023000` will be reduced to `035700230`

These zero sequences are much longer in reality. Almost all events are zero in fact

The cl includes a test that demonstrates this.


Bug: 301437959
Test: Added
Change-Id: I6da358cdbdd52632fa307c56dea426b617a2cd21