Enable Managed Devices for Emulator gRPC control
This moves the configuration of managed devices to the secondary stage
of UTP. This will enable emulator gRPC control for managed devices.

During the "beforeAll" step, when the device has been configured and
brought up, we will inspect the device and configure the token.

Once the emulator is configured an event will be fired with the optional
arguments that will be picked up by UTP.

Note: The dependency is not yet ready, so this does not yet work.

Test: Additional unit tests
Test: Local integration test (./gradlew device1DebugAndroidTest)
succeeds
Bug: 289283889
Change-Id: Ia8fe94c8ce88016f0a32b8fa3e02a9e1b36698f9

Change-Id: Ie18fe61bbcea644b8ca4dd96af21c454471b60db