Move resource-repository to JVM 17
Currently the default JDK for building the source code is 17. Specifying
custom JDK for building a target also applies restriction to the dependencies of the target. This adds additional maintenance burden and does not allow to use the new language features. Therefore, this commit changes JDK to the current default.

Bug: N/A
Test: N/A
Change-Id: I2d6da686ce99292e64f8ada4e651bbf03157be0d