Remove unnecessary assertion in DexArchiveBuilderTaskDelegate
The existing OutputMapping computation ensures that the dexing task runs
incrementally only if the jars' file paths are unchanged.

Given that condition, we then assert that the jars must not be reported
as REMOVED. However, that assertion may sometimes still fail
(see b/308401803#comment2 for possible explanations).

To address this issue, we will remove the assertion as it is not
strictly required. When the jar is reported as changed (regardless of
whether the change is REMOVED, MODIFIED, or ADDED), we will recompute
the jar's hash.

Bug: 308401803
Test: New test in DexArchiveBuilderDelegateTest
Change-Id: I5af7a7c66d65b35b272e7cfc67dff93868e6137b